% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_creation.R
\name{add_rows_to_holding_data}
\alias{add_rows_to_holding_data}
\title{Add a set of holding identifiers to a movenet holding data tibble}
\usage{
add_rows_to_holding_data(holding_data, holding_ids)
}
\arguments{
\item{holding_data}{A movenet-format holding data tibble (optional).}

\item{holding_ids}{A character vector with holding identifiers to add.}
}
\value{
A tibble like \code{holding_data}, but possibly with more rows. A warning
message is printed which lists any entries that have been added.
}
\description{
\code{add_rows_to_holding_data()} is an internal helper function that adds entries
to \code{holding_data}, using \code{holding_ids} as identifiers, and \code{NA} for any other
columns.
}
\keyword{internal}
