% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsening_functions.R
\name{anonymise}
\alias{anonymise}
\title{Anonymise data by replacing holding identifiers with prefix-integer
combinations}
\usage{
anonymise(data, prefix, key = NULL)
}
\arguments{
\item{data}{A holding or movement data frame.}

\item{prefix}{Character string, to form the basis of anonymised holding
identifiers. An integer will be appended to form this new identifier.}

\item{key}{A named list to be used as anonymisation key, or \code{NULL} (default)
to generate a new key. A provided \code{key} should have original holding
identifiers as names, and new (anonymised) identifiers as values.}
}
\value{
A named list with two elements:
\itemize{
\item \code{data} containing the anonymised data frame
\item \code{key} containing the applied anonymisation key. This has the form of a
named list, with original holding identifiers as names, and new (anonymised)
identifiers as values.
}
}
\description{
\code{anonymise()} anonymises a holding or movement data frame by replacing
holding identifiers with prefix-integer combinations. Both the anonymised
data frame and the anonymisation key are returned. By default, a new
anonymisation key is generated; alternatively, an existing key can be
provided.
}
\details{
Requires that the appropriate config file is loaded, to identify the
column(s) in \code{data} that contain(s) holding identifiers: origin (\code{from}) and
destination (\code{to}) columns for movement data, or the \code{id} column for holding
data.

If \code{key == NULL} (default), a new anonymisation key is generated, with
holdings being given new identifiers consisting of \code{prefix} followed by an
integer ranging between 1 and the total number of holdings. Integers are
assigned to holdings in a random order.

If an existing \code{key} is provided, its coverage of holding identifiers in
\code{data} is checked. If all holding identifiers in \code{data} are present among
element names in \code{key}, the \code{key} is used for anonymisation as-is: holding
identifiers in \code{data} are replaced with the values of elements of the same
name in \code{key}. Otherwise, if \code{data} contains holding identifiers that are not
present in \code{key}, the \code{key} is expanded by adding additional \code{prefix}-integer
combinations.
}
