% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_helpers.R
\name{are_ids_consec_intchars_from_1}
\alias{are_ids_consec_intchars_from_1}
\title{Are node ids consecutive "integers" (in character format), starting from 1?}
\usage{
are_ids_consec_intchars_from_1(node_ids)
}
\arguments{
\item{node_ids}{vector of node IDs (character format)}
}
\value{
a boolean - TRUE if the IDs need to be replaced, false otherwise
}
\description{
Checks whether node identifiers are consecutive "integers" (in character
format), ranging from 1 to the total number of node ids.
This is to facilitate conversion of movenet data tibbles into formats
required by networkDynamic and SimInf, which need node identifiers to be
consecutive integers starting from 1.
}
