% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_analysis_report.R
\name{basic_network_summary_per_time_unit}
\alias{basic_network_summary_per_time_unit}
\title{Title}
\usage{
basic_network_summary_per_time_unit(network, dates, time_unit, net_env)
}
\arguments{
\item{time_unit}{("month", "quarter", or "year")}
}
\description{
Title
}
