% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildNetwork}
\alias{buildNetwork}
\title{Builds a movement network from transfer data, incorporating length of stay at node as well as length 'away' from node.}
\usage{
buildNetwork(df, id, node.column, time.column, nodes.list, R_A, gamma,
  parallel = F, num.cores = parallel::detectCores() - 1)
}
\arguments{
\item{df}{Data frame of movement information. Must have a very specific structure, see \code{\link{Details}}}

\item{id}{Column containing individual identifier (unquoted)}

\item{node.column}{Column containing the node of the individual (unquoted)}

\item{time.column}{Column containing the time differences of the individual (unquoted).}

\item{nodes.list}{List containing all possible nodes (possibly larger than the list contained in df)}

\item{R_A}{Reproductive number of condition}

\item{gamma}{Rate of recovery of condition}

\item{parallel}{Compute in parallel (default \code{FALSE})}

\item{num.cores}{Number of cores to use in parallel computation (ignored if \code{parallel == FALSE})}
}
\value{
Adjacency matrix of the transfer network/graph.
}
\description{
Builds a movement network from transfer data, incorporating length of stay at node as well as length 'away' from node.
}
\details{
The structure of \code{df} is very specific. For a given individual, each stay will have 2 rows (in order):
the time of admission, and the time of discharge. The node will be the same for  these 2 rows.
The \code{time.column} will then have either the stay length at a given node (even rows), or the time to
next admission (odd rows). The first row of this column is expected to be NA.

The frame should be ordered by \code{id}, and then by the time (either admission or discharge). Sorting by \code{id}
is not required, but does make the frame nicer to look at.
}
