% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_config.R
\name{change_config}
\alias{change_config}
\title{Changing individual movenet configurations}
\usage{
change_config(...)
}
\arguments{
\item{...}{}
}
\description{
Changing individual movenet configurations
}
