% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsening_functions.R
\name{coarsen_date}
\alias{coarsen_date}
\title{Jitter and/or round down movement dates, and/or summarise data by time unit}
\usage{
coarsen_date(
  data,
  jitter,
  rounding_unit,
  week_start = getOption("lubridate.week.start", 7),
  sum_weight = TRUE,
  ...
)
}
\arguments{
\item{data}{A movement data frame.}

\item{jitter}{Either a positive integer, indicating the amount of jitter (in
days) to apply (see Details), or \code{FALSE} to not apply any jitter.}

\item{rounding_unit}{Either a character string specifying a time unit or a
multiple of a unit for movement dates to be rounded down to, or \code{FALSE} to
not apply any rounding. Valid base units are \code{day}, \code{week}, \code{month},
\code{bimonth}, \code{quarter}, \code{season}, \code{halfyear} and \code{year}. Arbitrary unique
English abbreviations as in the \code{\link[lubridate:period]{lubridate::period()}} constructor are
allowed. Rounding to multiples of units (except weeks) is supported.}

\item{week_start}{Week start day, only relevant for rounding down dates and/
or aggregating data by week. Default is the value of the
\code{lubridate.week.start} option, or 7 (Sunday) if this option is not set.
Full or abbreviated names of the days of the week can be in English or as
provided by the current locale.}

\item{sum_weight}{If this is \code{TRUE} (the default), weights are summed over
the specified \code{rounding_unit}, for all rows with the same origin and
destination. The name of the weight column will remain the same.}

\item{...}{Additional or
alternative summary function(s), of the form name = value, to pass on to
\code{\link[dplyr:summarise]{dplyr::summarise()}}. Any summary functions will be applied to \code{data},
grouped by origin, destination, and rounded-down date. The specified name
will be the column name in the resulting data frame.
The value can be:
\itemize{
\item A vector of length 1, e.g. \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.
\item A vector of length n, e.g. \code{quantile()}.
\item A data frame, to add multiple columns from a single expression.
}}
}
\value{
A movement data frame like \code{data}, but with jittered and/or rounded-down
movement dates.

If \code{sum_weight} is \code{TRUE} or any summary functions are provided through
\code{...}, the returned data frame contains data summarised by origin,
destination, and \code{rounding_unit}, and may thus have a decreased length
compared to \code{data}.

Columns for which a summary function is not provided, are dropped from
the resulting data frame.
}
\description{
\code{coarsen_date()} applies jitter to, and/or rounds down, movement dates in a
movement data frame. Rounding down is to the first day of the specified time
unit. It optionally summarises movement data with the same origin and
destination by this time unit. By default, this involves summation of
weights, but alternative or additional summary functions can be provided
through \code{...}.
}
\details{
Requires that the appropriate movement config file is loaded, to correctly
identify origin (\code{from}), destination (\code{to}), \code{date} and \code{weight} columns in
\code{data}.

If both jitter and rounding are applied, movement dates are first jittered
and then rounded down.

If \code{jitter > 0}, movement dates are modified by addition of a number of days
between \code{-jitter} and \code{jitter}, following a discrete uniform distribution. If
this were to result in a date being moved beyond the original date range, the
amount of jitter for this date is resampled, until the resulting date is
within the original date range.
If \code{jitter} is \code{FALSE} (or \code{jitter == 0}), no jitter is applied.

If \code{rounding_unit} is a character string, movement dates are rounded to the
first day of the \code{rounding_unit}.
If \code{rounding_unit} is \code{FALSE}, no rounding is applied.

If \code{rounding_unit} is a character string and \code{sum_weight} is \code{TRUE}, weights
are summed over all rows in \code{data} with the same origin, destination, and
rounded-down date. The summed weight column is temporarily renamed, so that
any additional weight-dependent summary functions passed through \code{...} use
the original weights rather than the sums.
If \code{rounding_unit} is \code{FALSE}, \code{sum_weight} is ignored.

If \code{rounding_unit} is a character string and any summary functions are
provided through \code{...}, the specified data are summarised accordingly, over
all rows in \code{data} with the same origin, destination, and rounded-down date.
Be careful when using existing names: the corresponding columns will be
immediately updated with the new data and this can affect subsequent
operations referring to this name.
If \code{rounding_unit} is \code{FALSE}, \code{...} is ignored.

Columns for which a summary function is not provided, are dropped from
the resulting data frame.
}
\seealso{
\code{\link[lubridate:round_date]{lubridate::floor_date()}}, \code{\link[dplyr:summarise]{dplyr::summarise()}} which this function
wraps.
}
