% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsening_functions.R
\name{coarsen_weight}
\alias{coarsen_weight}
\title{Jitter (add noise to) and/or round numeric movement data}
\usage{
coarsen_weight(
  data,
  column = movenetenv$options$movedata_cols$weight,
  jitter,
  round
)
}
\arguments{
\item{data}{A movement data frame.}

\item{column}{Name of a single numeric column to jitter and/or round. By
default this is the name of the weight column (as given in the movement
config file).}

\item{jitter}{Either a positive number, indicating the amount of jitter to
apply (see Details), or \code{FALSE} to not apply any jitter.}

\item{round}{Either a number, or \code{FALSE} to not apply any rounding. The data
in the selected column are rounded to the nearest multiple of this number.
\code{round} is additionally set as minimum possible value for the column.}
}
\value{
A movement data frame like \code{data}, but with jitter and/or rounding applied to
the selected numeric column.
}
\description{
\code{coarsen_weight()} applies jitter to, and/or rounds, a selected numeric
column in a movement data frame. By default, the weight column is selected.
}
\details{
Requires that the appropriate movement config file is loaded.

If both jitter and rounding are applied, the data in the selected column are
first jittered and then rounded.

If \code{jitter > 0}, the data in the selected column are modified by addition of
an amount of noise between \code{-jitter} and \code{-jitter}, following a uniform
distribution. If this were to result in a data point becoming \verb{<= 0}, the
amount of jitter for this data point is resampled, until the resulting data
point becomes positive. This is to capture that any movement in a livestock
movement database, is assumed to have a positive weight (quantity of animals
moved).
If \code{jitter == FALSE} (or \code{jitter == 0}), no jitter is applied.

If \code{round > 0}, the data in the selected column are modified by rounding to
multiples of \code{round}.
Additionally, any data points \verb{< round} are set to \code{round}, so that this
becomes the minimum possible value in the column. This is to capture that any
livestock movement, no matter how small, has an inherent risk that is
conceptually closer to that of \code{round}, than that of no movement at all.
If \code{round == FALSE} (or \code{round == 0}), no rounding is applied.
}
\seealso{
\code{\link[base:jitter]{base::jitter()}}, \code{\link[plyr:round_any]{plyr::round_any()}} which this function
wraps.
}
