% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_contact_chains.R
\name{contactchains2leaflet}
\alias{contactchains2leaflet}
\title{Create Leaflet map with ingoing and outgoing contact chains}
\usage{
contactchains2leaflet(
  movement_data,
  holding_data,
  admin_areas_map = NULL,
  colour_domain_weight_in = NULL,
  colour_domain_weight_out = NULL,
  colour_domain_moves_in = NULL,
  colour_domain_moves_out = NULL
)
}
\arguments{
\item{movement_data}{A movenet-format movement data tibble, with movements
that are part of ingoing and outgoing contact chains.}

\item{holding_data}{A movenet-format holding data tibble, with holdings that
are part of ingoing and outgoing contact chains.}

\item{admin_areas_map}{An sf object containing multipolygons for relevant
administrative areas.}

\item{colour_domain_moves_out}{}
}
\value{
A Leaflet map (HTML widget object) with graphics layers for holdings
and moves that are part of ingoing and outgoing contact chains, and for the
provided administrative area boundaries. Further layers can be added using
\verb{\%>\%}.
}
\description{
Internal helper function that plots ingoing and outgoing contact chains onto
a Leaflet map with default OpenStreetMap background.
}
\keyword{internal}
