% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_contact_chains.R
\name{ContactTrace2holdingdata}
\alias{ContactTrace2holdingdata}
\title{Convert ContactTrace object into movenet-format holding data}
\usage{
ContactTrace2holdingdata(contact_trace_object, original_holding_data = NULL)
}
\arguments{
\item{contact_trace_object}{A ContactTrace object returned by
EpiContactTrace's \code{Trace()} function.}

\item{original_holding_data}{An (optional) movenet-format holding data tibble
with additional data on the holdings in \code{contact_trace_object}.}
}
\value{
A movenet-format holding data tibble, with as rows only the
holdings that are part of the contact chains in \code{contact_trace_object}, and
as columns all original attribute columns from \code{original_holding_data} and
additionally "direction" ("root" for root holdings, "in" for holdings that
are part of ingoing contact chains, or "out" for moves that are part of
outgoing contact chains). If \code{original_holding_data} is \code{NULL}, a
tibble with only the columns "holding_id" and "direction" is returned.
}
\description{
Internal helper function that processes contact tracing results from
EpiContactTrace's \code{Trace()} function back into movenet format, and adds in
any holding attributes from the original movenet-format holding data tibble.
}
\keyword{internal}
