% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_contact_chains.R
\name{ContactTrace2movedata}
\alias{ContactTrace2movedata}
\title{Convert ContactTrace object into movenet-format movement data}
\usage{
ContactTrace2movedata(contact_trace_object, original_movement_data)
}
\arguments{
\item{contact_trace_object}{A ContactTrace object returned by
EpiContactTrace's \code{Trace()} function.}

\item{original_movement_data}{The movenet-format movement data tibble that was
used as input to \code{Trace()}.}
}
\value{
A movenet-format movement data tibble, with as rows only the
movements that are part of the contact chains in \code{contact_trace_object}, and
as columns all original attribute columns from \code{original_movement_data} and
additionally "direction" ("in" for moves that are part of ingoing contact
chains, or "out" for moves that are part of outgoing contact chains).
}
\description{
Internal helper function that processes contact tracing results from
EpiContactTrace's \code{Trace()} function back into movenet format, and adds in
any movement attributes from the original movenet-format movement data tibble.
}
\details{
This requires for "repeated movements on one day" to be aggregated,
otherwise it's complicated to ensure a correct join.
}
\seealso{
\code{\link[EpiContactTrace:Trace]{EpiContactTrace::Trace()}}
}
\keyword{internal}
