% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anonymisation_effect_analysis_report.R
\name{create_anonymisation_effect_analysis_report}
\alias{create_anonymisation_effect_analysis_report}
\title{Create a report analysing the effects of modifying movement weights on network
properties}
\usage{
create_anonymisation_effect_analysis_report(
  movement_data,
  output_file,
  modify_weights = TRUE,
  modify_dates = FALSE,
  n_jitter_sim = 3,
  time_unit = "28 days",
  data_reference = NULL,
  verbose = FALSE
)
}
\arguments{
\item{movement_data}{A movenet-format movement tibble.}

\item{output_file}{Output file name and path. If no path is provided, the
file will be saved in the current working directory.}

\item{modify_weights}{A logical indicating whether to analyse the effects of
modifying weights (default \code{TRUE}). Currently this MUST be \code{TRUE} for the
function to proceed.}

\item{modify_dates}{Does not currently do anything as analysis of modified
dates is not yet implemented (automatically set to \code{FALSE}).}

\item{n_jitter_sim}{An integer indicating the number of random jitter
simulations (default \code{3}).}

\item{time_unit}{A character string indicating the time unit of analyses
(default \code{"28 days"}). The overall movement data will be split into
sub-networks based on this time unit.}

\item{data_reference}{A character string providing an optional subtitle to
the report, e.g. an identifier for the original movement dataset (default
\code{NULL}).}

\item{verbose}{A logical indicating whether to print progress messages
(default \code{FALSE}).}
}
\value{
An html report file is created with text and visualisations
comparing the effects of movement weight modifications (different amounts
of jittering and rounding) on a selection of global network properties and
on the ranking of holdings according to several centrality measures.
}
\description{
\code{create_anonymisation_effect_analysis_report()} takes as input a livestock
movement dataset, and analyses how network analyses of these data are
affected by privacy-enhancing modifications (jittering or rounding) of
movement weights. The function creates an html report with visualisations of
the effects of different amounts of jittering and rounding on a selection of
epidemiologically relevant global network properties as well as on the
ranking of holdings according to various centrality measures. The aim is to
help livestock movement data managers and analysts find an appropriate
balance between data privacy and their utility in network analyses.
}
\details{
This function applies various privacy-enhancing modifications to the
provided \code{movement_data}, converts the resulting datasets into series of
static networks for each period of the specified \code{time_unit}, and then
provides users with a report with network analyses comparing the true data
with the modified data.

This function requires that an appropriate movement config file is loaded,
to correctly identify the weight column in \code{movement_data}.

Prior to applying privacy-enhancing modifications, \code{movement_data} is
pre-processed: repeated movements between the same holdings on the same day
are aggregated, and self-moves and moves with weight 0 are removed.

Privacy-enhancing modifications include the modification of movement
weights (batch sizes) by jittering and rounding. For jittering, the
function \code{\link[=jitter_weights]{jitter_weights()}} is used, with jitter ranges of 5, 10, 50, ...
up until the order of magnitude of the mean movement weight in the data.
For rounding, the function \code{\link[=round_weights]{round_weights()}} is used, with rounding units
of 5, 10, 50, ... up until the order of magnitude of the largest movement
weight in the data.

The periodic movement networks from true and modified networks are compared
with regards to two epidemiologically relevant weighted global network
properties: mean weighted shortest path length, and strength assortativity.
Additionally, comparisons are made with regards to the relative ranking of
holdings according to three weighted centrality measures: strength
(geometric mean of in- and out-strength), betweenness, and PageRank.
}
\examples{
# Set-up: Save movenet environment with current configurations
movenetenv <- movenet:::movenetenv
old_config <- movenetenv$options

# Load a movement config file
load_config(system.file("configurations", "ScotEID.yml",
                        package = "movenet"))

# Create a report analysing the effects of modifying movement weights
create_anonymisation_effect_analysis_report(example_movement_data,
                                            file.path(tempdir(), "mod_weight_analysis.html"),
                                            modify_weights = TRUE,
                                            modify_dates = FALSE,
                                            n_jitter_sim = 3,
                                            time_unit = "28 days",
                                            data_reference = "Example dataset",
                                            verbose = TRUE)

# Clean-up: Reinstate previous configurations and remove temporary objects
movenetenv$options <- old_config
rm("old_config", "movenetenv")
file.remove(file.path(tempdir(), "mod_weight_analysis.html"))


}
\seealso{
Other Privacy-enhancing functions: 
\code{\link{anonymise}()},
\code{\link{jitter_dates}()},
\code{\link{jitter_weights}()},
\code{\link{round_dates}()},
\code{\link{round_weights}()}
}
\concept{Network-related functions}
\concept{Privacy-enhancing functions}
