% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_creation.R
\name{create_edge_spells}
\alias{create_edge_spells}
\title{Create a networkDynamic edge.spells data frame from a movement data tibble}
\usage{
create_edge_spells(movement_data)
}
\arguments{
\item{movement_data}{A movement data tibble with integer node identifiers and
dates in numeric format.}
}
\value{
A networkDynamic-compliant \code{edge.spells} data frame, with 1 column
more than \code{movement_data}. The first 4 columns of the returned data frame
correspond to \code{date}, \code{date} again, \code{from}, and \code{to}; any additional
\code{movement_data} columns are located thereafter.
}
\description{
\code{create_edge_spells()} is an internal helper function that converts a
modified movenet-format movement data tibble into a networkDynamic-compliant
edge activity (\code{edge.spells}) data frame. It requires integer node
identifiers, and dates converted to numeric format.
}
\seealso{
\code{\link[networkDynamic:networkDynamic]{networkDynamic::networkDynamic()}} for \code{edge.spells} requirements
and how this data frame is used.
}
\keyword{internal}
