% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_modelling.R
\name{create_random_infected_vector}
\alias{create_random_infected_vector}
\title{Create a vector of holdings infected at t=0 to plug into SimInf models, by
randomly selecting a specified number of holdings}
\usage{
create_random_infected_vector(n_nodes, n_initially_infected)
}
\arguments{
\item{n_nodes}{Number of nodes/holdings included in the model.}

\item{n_initially_infected}{Number of holdings infected at t=0.}
}
\value{
A numeric vector, indicating which nodes are infected at t=0, to
plug into the \code{\link[siminf4movenet:SEIRcm]{siminf4movenet::SEIRcm()}} model or another SimInf model.
}
\description{
Create a vector of holdings infected at t=0 to plug into SimInf models, by
randomly selecting a specified number of holdings
}
\seealso{
\itemize{
\item \code{\link[siminf4movenet:SEIRcm]{siminf4movenet::SEIRcm()}} for the SEIRcm model developed for movenet.
\item \code{\link[SimInf:SimInf]{SimInf::SimInf}} for the underlying modelling package.
}

Other transmission modelling-related functions: 
\code{\link{SEIRcm_ASF}()},
\code{\link{create_specified_infected_vector}()},
\code{\link{data2contactmatrix}()},
\code{\link{get_cumulative_infected}()}
}
\concept{transmission modelling-related functions}
