% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a network analysis report from data}
\usage{
create_report(data, output, output_format = "html_document")
}
\arguments{
\item{data}{movement data}

\item{output}{output file name (and path)}

\item{output_format}{output format ("html_document", "pdf_document")}
}
\description{
Create a network analysis report from data
}
\keyword{internal}
