% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_analysis_report.R
\name{create_temporal_network_analysis_report}
\alias{create_temporal_network_analysis_report}
\title{Create a temporal network analysis report from data}
\usage{
create_temporal_network_analysis_report(
  network,
  whole_months = TRUE,
  time_unit = "month",
  output_file,
  output_format = "html_document"
)
}
\arguments{
\item{output_file}{output file name (and path)}

\item{output_format}{output format ("html_document", "pdf_document")}
}
\description{
Create a temporal network analysis report from data
}
