% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_helpers.R
\name{create_vertex_spells}
\alias{create_vertex_spells}
\title{Helper function: Create activity spells for nodes, to avoid
very slow reconcile.vertex.activity}
\usage{
create_vertex_spells(movement_data)
}
\arguments{
\item{movement_data}{The movement data tibble}
}
\value{
the vertex spells ready to be passed into networkDynamic()
}
\description{
Helper function: Create activity spells for nodes, to avoid
very slow reconcile.vertex.activity
}
