% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_admin_areas}
\alias{example_admin_areas}
\title{Example administrative area boundaries}
\format{
A sf simple feature collection (sf data frame) with 500 features (rows), 1
column, and the following properties:
\describe{
\item{Geometry type}{POLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -11.16411 ymin: 48.68521 xmax: 8.475314 ymax: 59.82427}
\item{CRS}{ETRS89 (EPSG:4258)}
\item{Features (columns)}{geometry: An sf POLYGON delineating the (fake) administrative area.}
}
}
\source{
The content of this dataset was generated by creating a boundary box and
Voronoi cells around the 500 points (synthetic holding coordinates) in
\code{movenet::example_holding_data}. The areas do not correspond to or derive
from any real administrative areas, nor do they relate to any real holding
locations.

The code used to generate this dataset is accessible at
\code{system.file("data-raw/example_admin_areas.R", package = "movenet")}.
}
\usage{
example_admin_areas
}
\description{
A dataset with synthetic administrative area boundaries, created to
illustrate the use of \code{trace_contact_chains()} with a map of area boundaries.
}
\seealso{
Other example data files and structures: 
\code{\link{example_holding_data.csv}},
\code{\link{example_holding_data}},
\code{\link{example_movement_data.csv}},
\code{\link{example_movement_data}},
\code{\link{local_spread_probabilities_ASF}}
}
\concept{example data files and structures}
\keyword{datasets}
