% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_holding_data.csv}
\alias{example_holding_data.csv}
\title{Example holding data file (delimited text)}
\format{
A UTF-8-encoded, comma-separated (csv) text file with 501 rows (incl. 1 header
row) and 7 columns:
\describe{
\item{cph}{Unique holding identifier; a character string in the format of
a UK County Parish Holding (CPH) number. This column corresponds to the
\code{id} data field, required in movenet-format holding data.}
\item{extra_column}{Numeric data column, using \code{.} as decimal mark.}
\item{easting, northing}{Holding coordinates (easting and northing) in an
all-numeric format (assumed to use the British National Grid projection).
These columns correspond to the \code{coord_x} and \code{coord_y} data fields,
optional in movenet-format holding data.}
\item{holding_type}{Type of holding (character string). This column
corresponds to the \code{type} data field, optional in movenet-format holding
data.}
\item{herd_size}{Herd size (numeric). This column corresponds to the
\code{herd_size} data field, optional in movenet-format holding data.}
}
}
\source{
The content of this dataset was generated by random number and text
generation, using UK-inspired data formats (e.g. County Parish Holding (CPH)
identifiers). It does not correspond to or derive from any real holding data.
}
\description{
A fully synthetic livestock holding data file, created to serve as example
input to movenet's \code{reformat_data()} function.
}
\seealso{
Other example data files and structures: 
\code{\link{example_holding_data}},
\code{\link{example_movement_data}},
\code{\link{example_movement_data.csv}},
\code{\link{local_spread_probabilities_ASF}}
}
\concept{example data files and structures}
