% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_holding_data}
\alias{example_holding_data}
\title{Example movenet-format holding data}
\format{
A data frame (tibble) with 500 rows and 4 columns:
\describe{
\item{cph}{Unique holding identifier; a character string in the format of
a UK County Parish Holding (CPH) number. This column corresponds to the
\code{id} data field, required in movenet-format holding data.}
\item{holding_type}{Type of holding (character string). This column
corresponds to the \code{type} data field, optional in movenet-format holding
data.}
\item{herd_size}{Herd size, in numeric (double) format. This column
corresponds to the \code{herd_size} data field, optional in movenet-format
holding data.}
\item{coordinates}{Holding coordinates in degrees longitude and latitude,
using the ETRS89 coordinate reference system, and formated as a simple
feature (sf) geometry list-column (class \code{sfc_POINT}). This column is
derived from \code{coord_x} and \code{coord_y} data fields, optional in
movenet-format holding data.}
}
}
\source{
The content of this dataset was generated by random number and text
generation, using UK-inspired data formats (e.g. County Parish Holding (CPH)
identifiers). It does not correspond to or derive from any real holding data.

The csv-format data file that this holding data tibble was derived from
is accessible at
\code{system.file("extdata/example_holding_data.csv", package = "movenet")}. The
tibble was obtained by calling \code{reformat_data()} on this csv file.
}
\usage{
example_holding_data
}
\description{
A fully synthetic livestock holding dataset, created to illustrate the
structure of movenet-format holding data tibbles.
}
\seealso{
\code{\link{vignette("movenet")}} for a detailed illustration of how this tibble
was obtained from \code{example_holding_data.csv}.

Other example data files and structures: 
\code{\link{example_holding_data.csv}},
\code{\link{example_movement_data.csv}},
\code{\link{example_movement_data}},
\code{\link{local_spread_probabilities_ASF}}
}
\concept{example data files and structures}
\keyword{datasets}
