% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_movement_data.csv}
\alias{example_movement_data.csv}
\title{Example movement data file (delimited text)}
\format{
A UTF-8-encoded, comma-separated (csv) text file with 5,001 rows (incl. 1
header row) and 19 columns:
\describe{
\item{movement_reference}{Unique numeric identifier of the movement.}
\item{foreign_reference, lot_no}{Additional identifiers (character strings).}
\item{lot_date}{Date at which \code{lot_no} was assigned, using DD/MM/YYYY
format (readr date format specification \verb{\%d/\%m/\%Y}).}
\item{departure_date, arrival_date}{Movement dates, using YYYYMMDD format
(readr date format specification \verb{\%Y\%m\%d}). Either column can be used as the
\code{date} data field, required in movenet-format movement data.}
\item{qty_pigs}{Movement weight/quantity (number of pigs moved). This
column corresponds to the \code{weight} data field, required in movenet-format
movement data.}
\item{qty_doa}{Number of pigs dead on arrival.}
\item{fci_declaration}{Food chain information declaration (numeric).}
\item{dep_assurance_no}{A numeric assurance identifier of the originating
keeper or holding.}
\item{dep_name, dep_address, dep_postcode}{Additional identifiers
(character strings) of the originating keeper or holding.}
\item{departure_cph}{Unique identifier of the originating holding; a
character string in the format of a UK County Parish Holding (CPH) number,
This column corresponds to the \code{from} data field, required in movenet-
format movement data.}
\item{dest_assurance_no}{A numeric assurance identifier of the destination
keeper or holding.}
\item{dest_name, dest_address, dest_postcode}{Additional identifiers
(character strings) of the destination keeper or holding.}
\item{dest_cph}{Unique identifier of the destination holding; a character
string in the format of a UK County Parish Holding (CPH) number. This column
corresponds to the \code{to} data field, required in movenet-format movement
data.}
}
}
\source{
The content of this dataset was generated by random number and text
generation, using UK-inspired data formats (e.g. County Parish Holding (CPH)
identifiers). It does not correspond to or derive from any real movement data.
}
\description{
A fully synthetic livestock movement data file, created to serve as example
input to movenet's \code{reformat_data()} function. It is inspired by and includes
data fields typically found in ScotEID's Scottish pig movement datasets.
}
\seealso{
Other example data files and structures: 
\code{\link{example_admin_areas}},
\code{\link{example_holding_data}},
\code{\link{example_holding_data.csv}},
\code{\link{example_movement_data}},
\code{\link{local_spread_probabilities_ASF}}
}
\concept{example data files and structures}
