% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_movement_data}
\alias{example_movement_data}
\title{Example movenet-format movement data}
\format{
A data frame (tibble) with 5,000 rows and 5 columns:
\describe{
\item{departure_cph}{Unique identifier of the holding where the movement
originates; a character string in the format of a UK County Parish Holding
(CPH) number. This column corresponds to the \code{from} data field, required in
movenet-format movement data.}
\item{dest_cph}{Unique identifier of the holding where the movement is
destined for; a character string in the format of a UK County Parish
Holding (CPH) number. This column corresponds to the \code{to} data field,
required in movenet-format movement data.}
\item{departure_date}{Movement date, using the R Date class. This column
corresponds to the \code{date} data field, required in movenet-format movement
data.}
\item{qty_pigs}{Movement weight/quantity (number of pigs moved), in numeric
(double) format. This column corresponds to the \code{weight} data field,
required in movenet-format movement data.}
\item{movement_reference}{A movement reference number. This column
illustrates an optional data field in movenet-format movement data.}
}
}
\source{
The content of this dataset was generated by random number and text
generation, using UK-inspired data formats (e.g. County Parish Holding (CPH)
identifiers). It does not correspond to or derive from any real movement data.

The csv-format data file that this movement data tibble was derived from
is accessible at
\code{system.file("extdata/example_movement_data.csv", package = "movenet")}. The
tibble was obtained by calling \code{reformat_data()} on this csv file.
}
\usage{
example_movement_data
}
\description{
A fully synthetic livestock movement dataset, created to illustrate the
structure of movenet-format movement data tibbles.
}
\seealso{
\code{\link{vignette("movenet")}} for a detailed illustration of how this tibble
was obtained from \code{example_movement_data.csv}.

Other example data files and structures: 
\code{\link{example_holding_data.csv}},
\code{\link{example_holding_data}},
\code{\link{example_movement_data.csv}},
\code{\link{local_spread_probabilities_ASF}}
}
\concept{example data files and structures}
\keyword{datasets}
