% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_analysis_report.R
\name{extract_dates}
\alias{extract_dates}
\title{Extract dates for study periods & subperiods (7d, 14d, 28d, 84d, month, quarter, year)}
\usage{
extract_dates(network, whole_months = TRUE)
}
\arguments{
\item{network}{A movement network}

\item{whole_months}{A logical indicating whether \code{movement_data} covers
full months (default \code{TRUE}). This affects calculation of average network
measures.}
}
\description{
Extract dates for study periods & subperiods (7d, 14d, 28d, 84d, month, quarter, year)
}
\keyword{internal}
