% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_helpers.R
\name{extract_monthly_networks}
\alias{extract_monthly_networks}
\title{Extract monthly movement networks}
\usage{
extract_monthly_networks(networks, n_threads, months_in_data)
}
\arguments{
\item{networks}{a named list of movement networks}

\item{months_in_data}{a list of start and end dates (in int format) for all months in the data}
}
\description{
Extract monthly movement networks
}
