% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_helpers.R
\name{extract_months}
\alias{extract_months}
\title{Extract months covered in movement dataset}
\usage{
extract_months(data)
}
\arguments{
\item{data}{Date column of movement data}
}
\value{
List of c(start date, end date) for each month covered in the dataset, with dates in int format
}
\description{
Extract months covered in movement dataset
}
