% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{extract_periodic_subnetworks}
\alias{extract_periodic_subnetworks}
\title{Extract periodic subnetworks from movement networks}
\usage{
extract_periodic_subnetworks(networks, n_threads, periods_in_data)
}
\arguments{
\item{networks}{A named list of movement networks.}

\item{n_threads}{A numeric indicating number of threads over which to
parallelise.}

\item{periods_in_data}{A list of start and end dates (in int format) for all periods in the data.}
}
\description{
Extract periodic subnetworks from movement networks
}
\details{
Note from network.extract: Note that only active vertices are
included by default (retain.all.vertices=FALSE). As a result, the size of
the extracted network may be smaller than the original. Vertex and edge
ids will be translated, but may not correspond to their original values.
If it is necessary to maintain the identities of vertices, see
persistent.ids. (Make this an argument? Add ... to pass on arguments?)
}
\keyword{internal}
