% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{extract_periods}
\alias{extract_periods}
\title{Extract time periods covered in movement dataset}
\usage{
extract_periods(data, period)
}
\arguments{
\item{data}{Date column of movement data}

\item{period}{time period for which to extract dates (n days, week, n weeks,
month, n months, year, n years)}
}
\value{
List of c(start date, end date) for each period covered in the data,
with dates in int format
}
\description{
Extract time periods covered in movement dataset
}
\details{
For periods of days or weeks, the first date in the data is used
as starting date; for periods of months or years, the first day of the
month, n months, year or n years is used as starting date. (OR make
starting date an argument?)
}
\keyword{internal}
