% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_creation.R
\name{filter_holding_data}
\alias{filter_holding_data}
\title{Filter a movenet holding data tibble for a set of holding identifiers}
\usage{
filter_holding_data(holding_data, holding_ids)
}
\arguments{
\item{holding_data}{A movenet-format holding data tibble (optional).}

\item{holding_ids}{A character vector with holding identifiers to keep.}
}
\value{
A tibble like \code{holding_data}, but possibly with fewer rows. A warning
message is printed which lists any entries that have been removed.
}
\description{
\code{filter_holding_data()} is an internal helper function that removes
\code{holding_data} entries of which the holding identifiers do not appear in
\code{holding_ids}.
}
\keyword{internal}
