% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations-related_functions.R
\name{get_config}
\alias{get_config}
\title{Get the values of configurations currently set in the movenet environment}
\usage{
get_config(...)
}
\arguments{
\item{...}{One or more configuration names, in the form of individual
character strings, a character vector, or a list.}
}
\value{
\itemize{
\item If no arguments are provided, an unstructured named list of all configurations
as currently set in the movenet environment.
\item If one or more configuration names are provided, an unstructured named list
with the current value of these configurations, or an empty named list and a
warning if the configurations are unset or ambiguous.
}
}
\description{
\code{get_config()} allows the user to examine the values of one or more
configurations as currently set in the movenet environment.
}
\details{
\subsection{Configuration names}{

Each configuration name must follow the \code{"datatype_configtype.configname"}
format, consisting of the following elements:
\itemize{
\item \code{datatype}: Type of data the configuration applies to. Either \code{movedata} or \code{holdingdata}.
\item \code{configtype}: Type of configuration. Either \code{fileopts} (file options) or \code{cols} (column headers/indices).
\item \code{configname}: Name of the specific configuration.
}

For example: \code{"movedata_fileopts.separator"}, \code{"holdingdata_cols.id"}

\code{get_config()} uses partial matching for configuration names. If a
configuration name is ambiguous or unrecognised (e.g. if the configuration is
unset), a warning is displayed.
}
}
\examples{
# Query the values of all configurations
get_config()

# Query the values of specific configurations
get_config("movedata_fileopts.separator", "holdingdata_cols.id")

}
\seealso{
\code{vignette("configurations")} for an explanation of the movenet
config system.

Other configurations-related functions: 
\code{\link{change_config}()},
\code{\link{load_config}()},
\code{\link{new_config}()},
\code{\link{save_config}()},
\code{\link{validate_config}()}
}
\concept{configurations-related functions}
