% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_helpers.R
\name{holdingids2consecints}
\alias{holdingids2consecints}
\title{Replaces node ids with consecutive "integers" (in character format), starting
from 1.}
\usage{
holdingids2consecints(
  movement_data,
  holding_data = NULL,
  incl_nonactive_holdings = FALSE
)
}
\arguments{
\item{movement_data}{Movenet format movement data tibble}

\item{holding_data}{Movenet format holding data tibble (optional)}

\item{incl_nonactive_holdings}{Whether to include holdings from
\code{holding_data} that are not present in \code{movement_data}. Default is \code{FALSE}.}
}
\value{
a named list with \code{key}, the modified \code{movement_data} tibble
\code{movement_data} tibble) and \code{holding_data} (the modified \code{holding_data}
tibble if present, or \code{NULL} otherwise)
}
\description{
Replaces node ids with consecutive "integers" (in character format), ranging
from 1 to the total number of node ids. This facilitates further conversion
of movenet data tibbles into formats required by networkDynamic and SimInf,
which need node identifiers to be consecutive integers starting from 1.
Also returns a "key" that links original ids and new "integer" ids, so that
original ids can be added as node attributes to networkDynamic and SimInf
outputs.
}
\keyword{internal}
