% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{infer_config_type}
\alias{infer_config_type}
\title{Infer the config type of a parsed yaml file}
\usage{
infer_config_type(yamlfile)
}
\arguments{
\item{yamlfile}{Named list, assumed to be a parsed yaml-format config file.}
}
\value{
Character vector representing config type. One of \code{"movement"},
\code{"holding"}, or \code{c("movement", "holding")}.
}
\description{
Internal helper function that detects any occurrence of \code{"move"} or \code{"holding"}
among \code{yamlfile}'s names, and infers from this the potential config type.
}
\keyword{internal}
