% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{internal_validate_config}
\alias{internal_validate_config}
\title{Validate a movenet config file and assemble "failed validation" messages}
\usage{
internal_validate_config(file, config_type = NULL)
}
\arguments{
\item{file}{Path of the config file to validate.}

\item{config_type}{The type of config file to expect: one of \code{NULL} (default,
resulting in inference of config type), \code{c("movement", "holding")}
(indicating a combined config file), \code{"movement"} or \code{"holding"}. Within the
movenet app this is used to indicate the \code{config_type} required by the app page.}
}
\value{
\itemize{
\item If the provided config file is valid, \code{NULL} (returned invisibly).
\item If the provided config file is not valid, a character vector (returned invisibly)
with one or more elements that each indicate a reason why validation has failed.
}
}
\description{
Internal helper function that runs a variety of validation functions on \code{file},
and collects and assembles any failed validation messages for \code{\link[=validate_config]{validate_config()}}.
\code{internal_validate_config()} validates that:
\itemize{
\item \code{file} can be parsed as a YAML string.
\item Any required/expected \code{config_type} is covered by the actual config type of \code{file}.
\item All required top-level, "fileopts", and "cols" configuration keys are present in \code{file}.
\item All configuration values match their data format requirements.
}
}
\details{
The movenet app uses this internal function instead of the external
\code{\link[=validate_config]{validate_config()}} function, to avoid raising app-crashing errors.

This internal function contains the \code{config_type} argument, whereas the
external \code{\link[=validate_config]{validate_config()}} function doesn't. This is because in the movenet
app, the movement and holding config files are provided via different pages,
requiring validation that the file's actual config type corresponds to the
config type of the app page. No such validation is required for standard
usage of \code{\link[=validate_config]{validate_config()}}via the package.
}
\seealso{
\itemize{
\item \code{vignette("configurations")} for an explanation of the movenet config system.
\item \code{list.files(system.file("configurations", package = "movenet"))} for a list of pre-installed, validated config files.
}
}
\keyword{internal}
