% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsening.R
\name{jitter_dates}
\alias{jitter_dates}
\title{Jitter (add noise to) movement dates}
\usage{
jitter_dates(data, range)
}
\arguments{
\item{data}{A movenet-format movement tibble.}

\item{range}{A positive integer, indicating the maximum amount of jitter (in
days) to apply to the movement dates (see Details).}
}
\value{
A movement tibble like \code{data}, but with jittered movement dates.
}
\description{
\code{jitter_dates()} adds random noise (jitter) to movement dates in a
movenet-format movement tibble.
}
\details{
Requires that an appropriate movement config file is loaded, to correctly
identify the \code{date} column in \code{data}.

Movement dates are modified by addition of a number of days between \code{-range}
and \code{range}, following a discrete uniform distribution. If this were to
result in a date being moved beyond the original date range, the amount of
jitter for this date is resampled, until the resulting date is within the
original date range.
}
