% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_config.R
\name{load_config}
\alias{load_config}
\title{Loading a movenet configurations file}
\usage{
load_config(configfile)
}
\arguments{
\item{configfile}{Either the path to a movenet configurations file, or the name
(minus extension) of a pre-installed configurations file.}
}
\description{
Loads configurations from a .yml format config file into the movenet
environment, so that movenet knows how to read the data files provided to it.
}
