% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{local_spread_probabilities_ASF}
\alias{local_spread_probabilities_ASF}
\title{ASF local spread transmission probabilities}
\format{
A data frame (tibble) with 5 rows and 3 columns:
\describe{
\item{lower_boundary}{Lower boundary of each distance band (inclusive),
in meters.}
\item{upper_boundary}{Upper boundary of each distance band (exclusive),
in meters.}
\item{probability}{Daily probability of infection through local spread,
given a distance from an infectious holding between \code{lower_boundary} and
\code{upper_boundary}.}
}
}
\source{
Halasa, T. \emph{et al.} (2016). Simulating the epidemiological and
economic effects of an African swine fever epidemic in industrialized swine
populations. \emph{Veterinary Microbiology}, \emph{193}, 7-16.
\url{https://doi.org/10.1016/j.vetmic.2016.08.004}
}
\usage{
local_spread_probabilities_ASF
}
\description{
Look-up table with daily probabilities of a susceptible holding becoming
infected with African swine fever through local spread, based on distance
from an infectious holding. Given as four distinct distance bands with
positive probabilities (indicating potential transmission) for distances up
to 2 km, and a band with probability 0 for distances >= 2 km. Local spread is
deemed to include, for example, transmission via rodents and insects,
unregistered animal movements, and sharing of equipment and tools.

Reference: Halasa, T. \emph{et al.} (2016). Simulating the epidemiological and
economic effects of an African swine fever epidemic in industrialized swine
populations. \emph{Veterinary Microbiology}, \emph{193}, 7-16.
\url{https://doi.org/10.1016/j.vetmic.2016.08.004}
}
\seealso{
Other example data files and structures: 
\code{\link{example_admin_areas}},
\code{\link{example_holding_data}},
\code{\link{example_holding_data.csv}},
\code{\link{example_movement_data}},
\code{\link{example_movement_data.csv}}
}
\concept{example data files and structures}
\keyword{datasets}
