% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_analysis_report.R
\name{monthxmonth_heatmap}
\alias{monthxmonth_heatmap}
\title{Create snapshot heatmap (e.g. loyalty analysis) for temporal network analysis and app}
\usage{
monthxmonth_heatmap(month_combinations, title)
}
\description{
Create snapshot heatmap (e.g. loyalty analysis) for temporal network analysis and app
}
\keyword{internal}
