% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_contact_chains.R
\name{movedata2EpiContactTrace}
\alias{movedata2EpiContactTrace}
\title{Make movement data compatible with EpiContactTrace's \code{Trace()} function}
\usage{
movedata2EpiContactTrace(movement_data)
}
\arguments{
\item{movement_data}{A movenet-format movement data tibble.}
}
\value{
A data.frame with columns "source" (= from), "destination" (= to),
"t" (= date), and "n" (= weight).
}
\description{
Internal helper function that processes movement data from a movenet-format
tibble to a format compatible with EpiContactTrace's \code{Trace()} function.
It selects the minimally required columns, renames them, and then converts
the tibble to a data.frame.
}
\seealso{
\code{\link[EpiContactTrace:Trace]{EpiContactTrace::Trace()}}
}
\keyword{internal}
