% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_creation.R
\name{movedata2igraph}
\alias{movedata2igraph}
\title{Create a static network representation of movenet-format movement data}
\usage{
movedata2igraph(
  movement_data,
  holding_data = NULL,
  incl_nonactive_holdings = FALSE
)
}
\arguments{
\item{movement_data}{A movenet-format movement data tibble.}

\item{holding_data}{A movenet-format holding data tibble (optional).}

\item{incl_nonactive_holdings}{A logical that indicates whether to include
holdings from \code{holding_data} that are not present in \code{movement_data}.
Default is \code{FALSE}. If set to \code{TRUE}, holdings that don't trade within the
period covered by \code{movement_data} are included in the network.}
}
\value{
An igraph object consisting of a directed static network, with nodes
representing holdings, and edges representing moves between holdings.
Additionally, information is printed out about the constructed network;
this output is passed on from \code{\link[igraph:graph_from_data_frame]{igraph::graph_from_data_frame()}}, the meaning
of all elements is explained in \code{\link[igraph:aaa-igraph-package]{igraph::igraph}}.
}
\description{
\code{movedata2igraph()} converts movenet-format movement and (optional)
holding data tibbles into an igraph static network representation. This
assumes the network is directed with no hyperedges. Loops are allowed;
multiplex edges can be made by using edge attributes.
}
\details{
Any \code{movement_data} columns beyond \code{from} and \code{to} are set as edge attributes
in the network, and any \code{holding_data} columns beyond \code{id} are set as node
attributes.
}
\examples{
movement_data <- head(example_movement_data, 20)
holding_data <- head(example_holding_data, 20)

# Create a network using only movement_data
movedata2igraph(movement_data)

# Create a network using movement_data and holding_data, including only
# active holdings
movedata2igraph(movement_data, holding_data,
                incl_nonactive_holdings = FALSE)

# Create a network using movement_data and holding_data, including both
# active and non-active holdings
movedata2igraph(movement_data, holding_data,
                incl_nonactive_holdings = TRUE)

rm(movement_data, holding_data)

}
\seealso{
\itemize{
\item \code{\link[igraph:graph_from_data_frame]{igraph::graph_from_data_frame()}} for the underlying network-generating process.
\item \code{\link[igraph:aaa-igraph-package]{igraph::igraph}} for an explanation of the printed output.
\item \code{vignette("network-analysis")} for an overview of the movenet network analysis workflow.
}

Other network-related functions: 
\code{\link{create_temporal_network_analysis_report}()},
\code{\link{movedata2networkDynamic}()},
\code{\link{parallel_summarise_temporal_node_properties}()},
\code{\link{trace_contact_chains}()}
}
\concept{network-related functions}
