% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_helpers.R
\name{movedata2networkDynamic}
\alias{movedata2networkDynamic}
\title{Create dynamic networks from movenet-format movement data}
\usage{
movedata2networkDynamic(
  movement_data,
  holding_data = NULL,
  incl_nonactive_holdings = FALSE
)
}
\arguments{
\item{movement_data}{Movenet format movement data}

\item{holding_data}{Movenet format holding data (optional)}

\item{incl_nonactive_holdings}{Whether to include holdings from
\code{holding_data} that are not present in \code{movement_data}. Default is \code{FALSE}.
If set to \code{TRUE}, holdings that don't trade within the period covered by
\code{movement_data} are added to the network but set as non-active.}
}
\value{
Temporal network (directed) in networkDynamic format
}
\description{
Create dynamic networks from movenet-format movement data
}
\details{
In the returned network, node identifiers are consecutive integers,
which may not correspond to original holding identifiers as provided in
\code{movement_data} and/or \code{holding_data}. The original holding identifiers (in
character format) have been set as 
\link[networkDynamic:persistent_ids]{networkDynamic::persistent.ids} and can be accessed through
\code{get.vertex.pid()}, or through the vertex attribute \code{true_id}.

In the returned network, holdings are set to be active only during movement
spells.
}
