% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movenet-package.R
\docType{package}
\name{movenet-package}
\alias{movenet}
\alias{movenet-package}
\title{movenet: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Authors:
\itemize{
  \item Carlijn Bogaardt \email{Carlijn.Bogaardt@glasgow.ac.uk}
  \item Stefan Widgren \email{stefan.widgren@sva.se} (\href{https://orcid.org/0000-0001-5745-2284}{ORCID})
}

}
\keyword{internal}
