% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations-related_functions.R
\name{new_config}
\alias{new_config}
\title{Copy a config file template to your working directory}
\usage{
new_config(config_type = c("movement", "holding"))
}
\arguments{
\item{config_type}{Character vector representing the type of config file for
which to copy the template: one of \code{c("movement", "holding")} (default,
indicating the combined config file template), \code{"movement"} or \code{"holding"}.}
}
\value{
No return value, but \code{new_config()} has the effect of writing a
YAML-format config file template to the current working directory. If this
is successful, a message is printed.
}
\description{
\code{new_config()} copies an empty config file template to the current working
directory.
}
\examples{
# Copy the movement config file template to your working directory
new_config(config_type = "movement")
readr::read_file(paste0(getwd(),"/movementconfig_template.yml")) #Examine file contents
# Clean-up: Remove the saved file
file.remove(paste0(getwd(),"/movementconfig_template.yml"))

}
\seealso{
\code{vignette("configurations")} for an explanation of the movenet
config system.

Other configurations-related functions: 
\code{\link{change_config}()},
\code{\link{get_config}()},
\code{\link{load_config}()},
\code{\link{save_config}()},
\code{\link{validate_config}()}
}
\concept{configurations-related functions}
