% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{parallel_max_reachabilities_with_id}
\alias{parallel_max_reachabilities_with_id}
\title{Extract max reachabilities and associated node persistent identifiers, in parallel}
\usage{
parallel_max_reachabilities_with_id(networks, n_threads)
}
\arguments{
\item{networks}{list of movement networks}

\item{n_threads}{}
}
\value{
list of lists: for each network, a list with (1) max reachability,
(2) the persistent identifiers of the node(s) with max reachability.
}
\description{
Extract max reachabilities and associated node persistent identifiers, in parallel
}
