% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{parallel_summarise_temporal_node_properties}
\alias{parallel_summarise_temporal_node_properties}
\title{Extract summary stats for temporal network node properties, in parallel}
\usage{
parallel_summarise_temporal_node_properties(
  networks,
  n_threads,
  node_property,
  statistics,
  identify_nodes = FALSE
)
}
\arguments{
\item{networks}{list of movement networks}

\item{n_threads}{number of threads over which to parallelise}

\item{node_property}{node property: one of c("forward reachability",
"temporal degree", "temporal indegree", "temporal outdegree")}

\item{statistics}{list with summary function(s) to calculate, e.g.
list(median, max = max).}

\item{identify_nodes}{whether you want to identify the nodes with maximal
values for \code{node_property} in each network. Default is \code{FALSE}. Requires
that max is included as a named function within \code{statistics}.}
}
\value{
named list of lists: for each network, a list consisting of:
(1) a named list \code{summary_statistics}, with the selected node property
summary statistic(s),
(2 - only if \code{identify_nodes} is set as \code{TRUE}) a named character vector
\code{node_pid_with_max_value}, with the persistent identifier(s) of the node(s)
with maximal value(s) for the selected node property.
}
\description{
Extract summary stats for temporal network node properties, in parallel
}
