% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_analysis_report.R
\name{periodic_data2summary_stats_df}
\alias{periodic_data2summary_stats_df}
\title{Create time-period based summary tables for temporal network analysis and app}
\usage{
periodic_data2summary_stats_df(
  periodic_data,
  period_dates,
  period = c("month", "quarter", "year")
)
}
\description{
Create time-period based summary tables for temporal network analysis and app
}
\keyword{internal}
