% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_helpers.R
\name{plot_measure_over_anonymisation_gradient}
\alias{plot_measure_over_anonymisation_gradient}
\title{Plot values of a measure for a varying range of jitter or rounding}
\usage{
plot_measure_over_anonymisation_gradient(data, measure_name, anonymisation)
}
\arguments{
\item{data}{tibble with measure values for range of jitter or rounding}

\item{measure_name}{measure name (to refer to in y axis)}

\item{anonymisation}{"jitter" or "round(ing)"}
}
\description{
Plot values of a measure for a varying range of jitter or rounding
}
