% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anonymisation_effect_analysis_report.R
\name{plot_measure_over_anonymisation_gradient2}
\alias{plot_measure_over_anonymisation_gradient2}
\title{Plot values of a measure for a varying range of jitter or rounding}
\usage{
plot_measure_over_anonymisation_gradient2(
  data,
  measure_name,
  anonymisation,
  plot_type,
  ...
)
}
\arguments{
\item{data}{tibble with measure values for range of jitter or rounding}

\item{measure_name}{measure name (to refer to in y axis)}

\item{anonymisation}{"jitter" or "round(ing)"}

\item{plot_type}{"boxplot" or "scatterplot"}

\item{...}{other parameters to pass on to wilcox.test}
}
\description{
Plot values of a measure for a varying range of jitter or rounding
}
\details{
Add alternative = "less" or alternative = "greater" for significance
stars from a one-sided test.
}
\keyword{internal}
