% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_analysis_report.R
\name{reachability_analysis}
\alias{reachability_analysis}
\title{Create reachability analysis content for temporal network report}
\usage{
reachability_analysis(network, dates, time_unit, net_env, n_threads)
}
\arguments{
\item{network}{networkDynamic object containing a movement network.}

\item{dates}{A named list with dates for study subperiods}

\item{time_unit}{Character string indicating the time unit of analyses. One
of \code{"month"}, \code{"quarter"}, or \code{"year"}.}

\item{net_env}{The internal environment within which plots are stored for the
report}

\item{n_threads}{An integer indicating the number of parallel threads to use
for reachability analysis.}
}
\description{
Create reachability analysis content for temporal network report
}
\keyword{internal}
