% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_data.R
\name{reformat_data}
\alias{reformat_data}
\title{Reading and reformatting of animal movement or holding data}
\usage{
reformat_data(datafile, type)
}
\arguments{
\item{datafile}{Path to a delimited file with movement or holding data.}

\item{type}{Data type: "movement" or "holding"}
}
\value{
reformatted movement or holding data (selected & renamed columns)
}
\description{
Reformats movement or holding data from a delimited file into a common
intermediate format (extracts and renames selected columns)
}
\examples{

}
