% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_data.R
\name{reformat_date}
\alias{reformat_date}
\title{Check date columns and change col type to date}
\usage{
reformat_date(date_col, date_format)
}
\arguments{
\item{date_format}{}
}
\description{
Check date columns and change col type to date
}
\keyword{internal}
