% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_data.R
\name{reformat_numeric}
\alias{reformat_numeric}
\title{Check numeric columns & change col type to numeric}
\usage{
reformat_numeric(numeric_col, decimal)
}
\arguments{
\item{decimal}{}
}
\description{
Check numeric columns & change col type to numeric
}
\keyword{internal}
