% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsening_functions.R
\name{replace_ids_w_key}
\alias{replace_ids_w_key}
\title{Replace holding identifiers using a key}
\usage{
replace_ids_w_key(data, col_to_anonymise, key)
}
\arguments{
\item{data}{Dataframe or tibble (e.g. movement data)}

\item{col_to_anonymise}{Headers or indices of columns in \code{data} that
contain the holding identifiers to be replaced}

\item{key}{Named character vector, with original holding identifiers as names
and new identifiers as values.}
}
\value{
A modified version of \code{data} where, within columns indicated by
\code{col_to_anonymise}, holding identifers have been replaced according to \code{key}.
}
\description{
\code{replace_ids_w_key()} replaces holding identifiers according to a user-
provided key
}
\keyword{internal}
