% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsening_functions.R
\name{round_dates}
\alias{round_dates}
\title{Round down movement dates, and/or summarise data by time unit}
\usage{
round_dates(
  data,
  unit,
  week_start = getOption("lubridate.week.start", 7),
  sum_weight = TRUE,
  ...
)
}
\arguments{
\item{data}{A movenet-format movement tibble.}

\item{unit}{A character string specifying a time unit or a multiple of a
unit for movement dates to be rounded down to. Valid base units are \code{day},
\code{week}, \code{month}, \code{bimonth}, \code{quarter}, \code{season}, \code{halfyear} and \code{year}.
Arbitrary unique English abbreviations as in the \code{\link[lubridate:period]{lubridate::period()}}
constructor are allowed. Rounding to multiples of units (except weeks) is
supported.}

\item{week_start}{Week start day, only relevant for rounding down dates and/
or aggregating data by week. Default is the value of the
\code{lubridate.week.start} option, or 7 (Sunday) if this option is not set.
Full or abbreviated names of the days of the week can be in English or as
provided by the current locale.}

\item{sum_weight}{If this is \code{TRUE} (the default), weights are summed over
\code{unit}, for all rows with the same origin and destination. The name of
the weight column will remain the same.}

\item{...}{\code{\link[dplyr:dplyr_data_masking]{data-masking}} Additional or
alternative summary function(s), of the form name = value, to pass on to
\code{\link[dplyr:summarise]{dplyr::summarise()}}. Any summary functions will be applied to \code{data},
grouped by origin, destination, and rounded-down date. The specified name
will be the column name in the resulting tibble.
The value can be:
\itemize{
\item A vector of length 1, e.g. \code{min(x)}, \code{n()}, or \code{sum(is.na(y))}.
\item A vector of length n, e.g. \code{quantile()}.
\item A tibble, to add multiple columns from a single expression.
}}
}
\value{
A movement tibble like \code{data}, but with rounded-down movement dates.

If \code{sum_weight} is \code{TRUE} or any summary functions are provided through
\code{...}, the returned tibble contains data summarised by origin,
destination, and \code{unit}, and may thus have a decreased length compared to
\code{data}.

Columns for which a summary function is not provided, are dropped from
the resulting tibble.
}
\description{
\code{round_dates()} rounds movement dates in a movenet-format movement
tibble, down to the first day of the specified time unit. It optionally
summarises movement data with the same origin and destination by this time
unit. By default, this involves summation of weights, but alternative or
additional summary functions can be provided through \code{...}.
}
\details{
Requires that an appropriate movement config file is loaded, to correctly
identify origin (\code{from}), destination (\code{to}), \code{date} and \code{weight} columns in
\code{data}.

Movement dates are rounded to the first day of the rounding unit specified in
\code{unit}.

If \code{sum_weight} is \code{TRUE}, weights are summed over all rows in \code{data} with
the same origin, destination, and rounded-down date. The summed weight column
is temporarily renamed, so that any additional weight-dependent summary
functions passed through \code{...} use the original weights rather than the sums.

If any summary functions are provided through \code{...}, the specified data are
summarised accordingly, over all rows in \code{data} with the same origin,
destination, and rounded-down date. Be careful when using' existing names:
the corresponding columns will be immediately updated with the new data and
this can affect subsequent operations referring to this name.

Columns for which a summary function is not provided, are dropped from
the resulting tibble.
}
\seealso{
\code{\link[lubridate:round_date]{lubridate::floor_date()}}, \code{\link[dplyr:summarise]{dplyr::summarise()}} which this function
wraps.
}
