% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsening_functions.R
\name{round_weights}
\alias{round_weights}
\title{Round numeric movement data}
\usage{
round_weights(data, unit, column = movenetenv$options$movedata_cols$weight)
}
\arguments{
\item{data}{A movenet-format movement tibble.}

\item{unit}{A positive number. The data in the selected column are rounded to
the nearest multiple of this number. \code{unit} is additionally set as minimum
possible value for the column.}

\item{column}{Name of a single numeric column to round. By default this is
the weight column (as specified in the loaded movement configurations).}
}
\value{
A movement tibble like \code{data}, but with rounding applied to the selected
numeric column.
}
\description{
\code{round_weights()} rounds a selected numeric column in a movenet-format
movement tibble. By default, the weight column is selected.
}
\details{
Requires that the appropriate movement config file is loaded, to correctly
identify the \code{weight} column in \code{data}.

The data in the selected column are modified by rounding to multiples of
\code{unit}.
Additionally, any data points \verb{< unit} are set to \code{unit}, so that this
becomes the minimum possible value in the column. This is to capture that any
livestock movement, no matter how small, has an inherent risk that is
conceptually closer to that of \code{unit}, than that of no movement at all.
}
\examples{
# Set-up: Save movenet environment with current configurations
movenetenv <- movenet:::movenetenv
old_config <- movenetenv$options

# Load a movement config file
load_config(system.file("configurations", "ScotEID.yml",
                        package = "movenet"))

# Round movement weights to multiples of 5 (and set 5 as minimum value)
moves_w_rounded_weights <- round_weights(example_movement_data, 5)
head(moves_w_rounded_weights)

# Round movement reference numbers to multiples of 10 (and set 10 as minimum
# value)
moves_w_rounded_refs <-
  round_weights(example_movement_data, 10, column = "movement_reference")
head(moves_w_rounded_refs)

# Clean-up: Reinstate previous configurations
movenetenv$options <- old_config
rm("old_config", "movenetenv", "moves_w_rounded_weights",
   "moves_w_rounded_refs")
}
\seealso{
\code{\link[plyr:round_any]{plyr::round_any()}}, which this function wraps.

Other Privacy-enhancing functions: 
\code{\link{anonymise}()},
\code{\link{create_anonymisation_effect_analysis_report}()},
\code{\link{jitter_dates}()},
\code{\link{jitter_weights}()},
\code{\link{round_dates}()}
}
\concept{Privacy-enhancing functions}
