% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsening_functions.R
\name{round_weights_signif}
\alias{round_weights_signif}
\title{Round numeric movement data to a specified number of significant digits}
\usage{
round_weights_signif(
  data,
  digits,
  column = movenetenv$options$movedata_cols$weight
)
}
\arguments{
\item{data}{A movenet-format movement tibble.}

\item{digits}{An integer indicating the number of significant digits to be
used for rounding. Accepted values are between 1 and 22. \code{digits} also
determines the minimum possible value for the column (set at \code{10^digits}).}

\item{column}{Name of a single numeric column to round. By default this is
the weight column (as specified in the loaded movement configurations).}
}
\value{
A movement tibble like \code{data}, but with rounding applied to the selected
numeric column.
}
\description{
\code{round_weights_signif()} rounds a selected numeric column in a movenet-format
movement tibble to the specified number of significant digits. By default,
the weight column is selected.
}
\details{
Requires that the appropriate movement config file is loaded, to correctly
identify the \code{weight} column in \code{data}.

The data in the selected column are modified by rounding to the number of
significant digits specified by \code{digits}.
Additionally, any data points with value \verb{< 10^digits} are set to \code{10^digits},
so that this becomes the minimum possible value in the column.
}
\seealso{
\code{\link[=signif]{signif()}}, which this function wraps.
}
