% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRcm.R
\name{SEIRcm}
\alias{SEIRcm}
\title{Create an SEIR model with contact matrix}
\usage{
SEIRcm(u0, tspan, events = NULL, beta = NULL, epsilon = NULL, gamma = NULL)
}
\arguments{
\item{u0}{FIXME}

\item{tspan}{FIXME}

\item{events}{FIXME}

\item{beta}{FIXME}

\item{epsilon}{FIXME}

\item{gamma}{FIXME}
}
\description{
Create an SEIR model with contact matrix
}
