% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_modelling.R
\name{SEIRcm_ASF}
\alias{SEIRcm_ASF}
\title{Create an SEIR model with contact matrix and ASF-inspired parameters}
\usage{
SEIRcm_ASF(infected, tspan, contact_matrix)
}
\arguments{
\item{infected}{logical vector with the infectious status in each
node.}

\item{tspan}{A vector (length >= 1) of increasing time points
where the state of each node is to be returned. Can be either
an \code{integer} or a \code{Date} vector. A \code{Date}
vector is coerced to a numeric vector as days, where
\code{tspan[1]} becomes the day of the year of the first year
of \code{tspan}. The dates are added as names to the numeric
vector.}

\item{contact_matrix}{an optional contact matrix to include
between-node transmission of infection. The contact matrix
specifies the strength of the interaction between
nodes. Specifically, \code{contact_matrix[i, j]} measures the
relative strength of transmission from node \code{i} to node
\code{j}.}
}
\description{
Creates a SimInf \code{SEIRcm} (\code{\link[siminf4movenet:SEIRcm]{siminf4movenet::SEIRcm()}}) model with the
following fixed parameters relating to African swine fever:
\itemize{
\item \code{epsilon_rate = 18/6.1}
\item \code{epsilon_shape = 18}
\item \code{gamma_rate = 1/8}
\item \code{gamma_shape = 1}
}

N.B. This function requires the siminf4movenet package to be installed.
}
\details{
\subsection{Holding-level incubation rate (epsilon) related parameters}{

Parameters for Epsilon come from \href{https://doi.org/10.1111/tbed.12748}{Guinat et al. 2018},
via \href{https://doi.org/10.2903/j.efsa.2020.5996}{EFSA 2020}.
They are within-farm transmission parameters for one of 9 outbreaks of the
Georgia 2007/1 strain in Russia. The rate and shape parameters are both close
to the median for the 9 outbreaks. In the EFSA report, parameters from this
outbreak are used as a low-R0 scenario.
These parameters cover a larger range with a higher average, compared to
several other studies - but this might fit with the idea that farm-level
latent periods would be higher than individual-pig latent periods due to v low
level transmissibility if considering just a single pig.
}

\subsection{Holding-level removal rate (gamma) related parameters}{

Average time-to-removal (infectious period) was derived as follows:
average time from first infection to suspicion report = 13 days (\href{https://doi.org/10.2903/j.efsa.2020.5996}{EFSA 2020}) +
assumed time from suspicion report to removal = 1 day -
average latent period = 6 days (as from \href{https://doi.org/10.1111/tbed.12748}{Guinat et al. 2018})
}
}
\seealso{
\itemize{
\item \code{\link[siminf4movenet:SEIRcm]{siminf4movenet::SEIRcm()}} for the underlying SEIRcm model
\item \code{\link[SimInf:SimInf]{SimInf::SimInf}} for the underlying modelling package
}

Other transmission modelling-related functions: 
\code{\link{create_random_infected_vector}()},
\code{\link{create_specified_infected_vector}()},
\code{\link{data2contactmatrix}()},
\code{\link{get_cumulative_infected}()}
}
\concept{transmission modelling-related functions}
