% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anonymisation_effect_analysis_report.R
\name{series_5_10_50_100_up_to_oom}
\alias{series_5_10_50_100_up_to_oom}
\title{Get series of 5, 10, 50, 100, ... until order of magnitude of x}
\usage{
series_5_10_50_100_up_to_oom(x)
}
\description{
Get series of 5, 10, 50, 100, ... until order of magnitude of x
}
\keyword{internal}
