% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_analysis_report.R
\name{snapshot_boxplot}
\alias{snapshot_boxplot}
\title{Create snapshot boxplots (e.g. reachability analyses) for temporal network analysis and app}
\usage{
snapshot_boxplot(periodic_data, title, ylab)
}
\description{
Create snapshot boxplots (e.g. reachability analyses) for temporal network analysis and app
}
\keyword{internal}
