% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_analysis_report.R
\name{snapshot_summary_stats_linechart}
\alias{snapshot_summary_stats_linechart}
\title{Create snapshot summary linechart (e.g. batchsize plot) for temporal network analysis and app}
\usage{
snapshot_summary_stats_linechart(periodic_data, title, ylab)
}
\description{
Create snapshot summary linechart (e.g. batchsize plot) for temporal network analysis and app
}
\keyword{internal}
