% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_analysis_report.R
\name{snapshots_diffperiods_dotplot1}
\alias{snapshots_diffperiods_dotplot1}
\title{Create snapshot dotplots for temporal network report and app}
\usage{
snapshots_diffperiods_dotplot1(
  data_daily,
  data_7days,
  data_14days,
  data_28days,
  data_84days,
  title,
  ylab
)
}
\description{
Create snapshot dotplots for temporal network report and app
}
\keyword{internal}
