% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{split_igraph}
\alias{split_igraph}
\title{Split up a movenet movement igraph by movement date}
\usage{
split_igraph(g, period, incl_nonactive_holdings = FALSE)
}
\arguments{
\item{g}{An igraph object made from movenet movement (and optional holding) data.}

\item{period}{time period for which to extract dates (n days, week, n weeks,
month, n months, year, n years)}

\item{incl_nonactive_holdings}{A logical that indicates whether to keep
holdings that don't trade within the relevant time \code{period}. Default is
\code{FALSE}.}
}
\value{
A list of igraph objects for each \code{period} in the data.
}
\description{
\code{split_movement_data()} is an internal helper function that splits a movenet
movement data containing igraph up into separate igraphs for each time
\code{period}.
}
\details{
Requires that an appropriate movement config file is loaded, to correctly
identify the \code{date} attribute in \code{g}.

For periods of days or weeks, the first date in the data is used
as starting date; for periods of months or years, the first day of the
month, n months, year or n years is used as starting date. (OR make
starting date an argument?)
}
\keyword{internal}
