% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{split_movement_data}
\alias{split_movement_data}
\title{Split up a movenet movement data tibble by movement date}
\usage{
split_movement_data(movement_data, period)
}
\arguments{
\item{movement_data}{A movenet-format movement data tibble.}

\item{period}{time period for which to extract dates (n days, week, n weeks,
month, n months, year, n years)}
}
\value{
A list of movenet-format movement data tibbles, for each \code{period} in the data.
}
\description{
\code{split_movement_data()} is an internal helper function that splits
\code{movement_data} up into separate tibbles for each time \code{period}.
}
\details{
Requires that an appropriate movement config file is loaded, to correctly
identify the \code{date} column in \code{movement_data}.

For periods of days or weeks, the first date in the data is used
as starting date; for periods of months or years, the first day of the
month, n months, year or n years is used as starting date. (OR make
starting date an argument?)
}
\keyword{internal}
