% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_analysis_report.R
\name{static_network_snapshots_analysis}
\alias{static_network_snapshots_analysis}
\title{Create component and loyalty analyses content for temporal network report and app}
\usage{
static_network_snapshots_analysis(network, dates, time_unit, net_env)
}
\arguments{
\item{network}{networkDynamic object containing a movement network.}

\item{dates}{A named list with dates for study subperiods}

\item{time_unit}{Character string indicating the time unit of analyses. One
of \code{"month"}, \code{"quarter"}, or \code{"year"}.}

\item{net_env}{The internal environment within which plots are stored for the
report}
}
\description{
Create component and loyalty analyses content for temporal network report and app
}
\keyword{internal}
