% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_contact_chains.R
\name{trace_contact_chains}
\alias{trace_contact_chains}
\title{Trace ingoing and outgoing contact chains and visualise data on a Leaflet map}
\usage{
trace_contact_chains(
  movement_data,
  holding_data,
  root,
  tEnd = NULL,
  days = NULL,
  inBegin = NULL,
  inEnd = NULL,
  outBegin = NULL,
  outEnd = NULL,
  maxDistance = NULL,
  admin_areas_map = NULL,
  colour_domain_weight_in = NULL,
  colour_domain_weight_out = NULL,
  colour_domain_moves_in = NULL,
  colour_domain_moves_out = NULL
)
}
\arguments{
\item{movement_data}{A movenet-format movement tibble.}

\item{holding_data}{A movenet-format holding tibble that includes a
"coordinates" column.}

\item{root}{A vector of root holdings to perform contact tracing for.}

\item{tEnd}{the last date to include ingoing and outgoing
movements. Defaults to \code{NULL}}

\item{days}{the number of previous days before tEnd to include
ingoing and outgoing movements. Defaults to \code{NULL}}

\item{inBegin}{the first date to include ingoing
movements. Defaults to \code{NULL}}

\item{inEnd}{the last date to include ingoing movements. Defaults
to \code{NULL}}

\item{outBegin}{the first date to include outgoing
movements. Defaults to \code{NULL}}

\item{outEnd}{the last date to include outgoing
movements. Defaults to \code{NULL}}

\item{maxDistance}{stop contact tracing at maxDistance (inclusive)
from root. Default is \code{NULL} i.e. don't use the
maxDistance stop criteria.}

\item{admin_areas_map}{An sf dataframe with administrative area boundaries.}

\item{colour_domain_weight_in}{A numeric vector representing the range (i.e.
\code{c(min, max)}) of possible aggregate ingoing weight values to colour
administrative areas by. If \code{NULL} (default), the range will be determined
from the data.}

\item{colour_domain_weight_out}{A numeric vector representing the range (i.e.
\code{c(min, max)}) of possible aggregate outgoing weight values to colour
administrative areas by.  If \code{NULL} (default), the range will be determined
from the data.}

\item{colour_domain_moves_in}{A numeric vector representing the range (i.e.
\code{c(min, max)}) of possible aggregate numbers of ingoing moves to colour
administrative areas by. If \code{NULL} (default), the range will be determined
from the data.}

\item{colour_domain_moves_out}{A numeric vector representing the range (i.e.
\code{c(min, max)}) of possible aggregate numbers of outgoing moves to colour
administrative areas by. If \code{NULL} (default), the range will be determined
from the data.}
}
\value{
A Leaflet map (HTML widget object) with graphics layers for holdings
and movements that are part of ingoing and outgoing contact chains, and for
administrative area data if \code{admin_areas_map} is provided. Further layers
can be added using \verb{\%>\%}.
}
\description{
\code{trace_contact_chains()} determines ingoing and outgoing contact chains for
one or more \code{root} holdings over a specified time period, and then plots
these on an interactive Leaflet map. If a map with administrative areas is
provided, contact chain data are aggregated geographically, and the resulting
map can be coloured by the total weight or number of ingoing movements from
or outgoing movements to each area.
}
\details{
The time period can either be specified using \code{tEnd} and \code{days}, or
using \code{inBegin}, \code{inEnd}, \code{outBegin} and \code{outEnd}. If using \code{tEnd} and
\code{days}, the time period for ingoing and outgoing contacts ends at \code{tEnd} and
starts at \code{days} prior to \code{tEnd}. The tracing will be performed for each
combination of \code{root}, \code{tEnd} and \code{days}. An alternative way is to use
\code{inBegin}, \code{inEnd}, \code{outBegin} and \code{outEnd}. The time period for ingoing
contacts starts at \code{inBegin} and ends at \code{inEnd}. For outgoing contacts the
time period starts at \code{outBegin} and ends at \code{outEnd}. The vectors \code{root},
\code{inBegin}, \code{inEnd}, \code{outBegin}, and \code{outEnd} must have the same lengths and
the tracing will be performed for each index of them.

If \code{admin_areas_map} is provided, administrative area boundaries are added to
the Leaflet map. Additionally, contact chain data are summarised
geographically and the user can choose to colour the map by the total weight
or number of ingoing movements from or outgoing movements to each
administrative area. By default, the colour scales for these geographical
layers are determined by the ranges of the respective values in the data,
however they can be set via the \code{colour_domain_weight_in},
\code{colour_domain_weight_out}, \code{colour_domain_moves_in}, and/or
\code{colour_domain_moves_out} arguments in order to make colours comparable
across maps.
}
\seealso{
\itemize{
\item \code{\link[EpiContactTrace:Trace]{EpiContactTrace::Trace()}} for the underlying contact chain tracing functionality.
\item \code{\link[leaflet:leaflet]{leaflet::leaflet()}} for the underlying mapping functionality.
}
}
