% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_config}
\alias{validate_config}
\title{Validate a movenet config file}
\usage{
validate_config(file)
}
\arguments{
\item{file}{config file (yml format) to validate}
}
\description{
Validate a movenet config file
}
