% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_config_cols}
\alias{validate_config_cols}
\title{Validate that required "cols" keys are present in a parsed yaml file}
\usage{
validate_config_cols(yamlfile, config_type)
}
\arguments{
\item{yamlfile}{Named list, assumed to be a parsed yaml-format config file.}

\item{config_type}{Character vector representing the type of configurations
for which to check the "cols" keys: one of \code{c("movement", "holding")},
\code{"movement"} or \code{"holding"}.}
}
\value{
Either \code{NULL} if all "cols" (column header/index) keys required for \code{config_type}
are present, or a character string stating that \code{file} has an unexpected
structure and which "cols" keys are missing.
}
\description{
Internal helper function that checks:
\itemize{
\item Whether all "cols" (column header/index) keys required for \code{config_type} are present in \code{yamlfile}.
\item Whether all configurations within \code{movedata_cols} or \code{holdingdata_cols} have unique values.
\item In case any geographical coordinate-related "cols" configurations (\code{holdingdata_cols.coord_x}
and \code{holdingdata_cols.coord_y}) are included, that both of these are present.
A character string containing a failed validation message is printed in case
any keys are missing or any values are duplicated.
}
}
\details{
Required "cols" keys for \code{config_type = "movement"} are as follows:
\itemize{
\item \code{movedata_cols.from}
\item \code{movedata_cols.to}
\item \code{movedata_cols.date}
\item \code{movedata_cols.weight}
}

Required "cols" keys for \code{config_type = "holding"} are as follows:
\itemize{
\item \code{holdingdata_cols.id}
\item For the inclusion of geographical coordinates in any data extraction and
analysis, both \code{holdingdata_cols.coord_x} and \code{holdingdata_cols.coord_y} are required.
}

All configurations within \code{movedata_cols} or \code{holdingdata_cols} must have unique values.
}
\keyword{internal}
