% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_config_fileopts}
\alias{validate_config_fileopts}
\title{Validate that required "fileopts" keys are present in a parsed yaml file}
\usage{
validate_config_fileopts(yamlfile, config_type)
}
\arguments{
\item{yamlfile}{Named list, assumed to be a parsed yaml-format config file.}

\item{config_type}{Character vector representing the type of configurations
for which to check the "fileopts" keys: one of \code{c("movement", "holding")},
\code{"movement"} or \code{"holding"}.}
}
\value{
Either \code{NULL} if all "fileopts" (file options) keys required for \code{config_type}
are present, or a character string stating that \code{file} has an unexpected
structure and which "fileopts" keys are missing.
}
\description{
Internal helper function that checks:
\itemize{
\item Whether all "fileopts" (file options) keys required for \code{config_type} are present in \code{yamlfile}.
\item In case any geographical coordinate-related "cols" configurations (\code{holdingdata_cols.coord_x}
and \code{holdingdata_cols.coord_y}) are included, that all "fileopts" keys required
for geographic data are present in \code{yamlfile}.
A character string containing a failed validation message is printed in case
any keys are missing.
}
}
\details{
Required "fileopts" keys for \code{config_type = "movement"} are as follows:
\itemize{
\item \code{movedata_fileopt.separator}
\item \code{movedata_fileopts.decimal}
\item \code{movedata_fileopts.encoding}
\item \code{movedata_fileopts.date_format}
}

Required "fileopts" keys for \code{config_type = "holding"} are as follows:
\itemize{
\item \code{holdingdata_fileopt.separator}
\item \code{holdingdata_fileopts.decimal}
\item \code{holdingdata_fileopts.encoding}
\item For the inclusion of geographical coordinates in any data extraction and
analysis, both \code{holdingdata_fileopts.country_code} and
\code{holdingdata_fileopts.coord_EPSG_code} are required.
}
}
\keyword{internal}
