% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_config_root}
\alias{validate_config_root}
\title{Validate that top-level config keys are present in a parsed yaml file}
\usage{
validate_config_root(yamlfile)
}
\arguments{
\item{yamlfile}{Named list, assumed to be a parsed yaml-format config file.}
}
\value{
Either \code{NULL} if all movement and/or holding top-level config keys are present,
or a character string stating that \code{file} has an unexpected structure and which
top-level config keys are missing.
}
\description{
Internal helper function that checks whether all movement and/or holding
top-level config keys are present in \code{yamlfile}, and prints a character string
containing a failed validation message in case any such keys are missing.
}
\details{
Required top-level config keys are as follows:
\itemize{
\item \code{movedata_fileopts} and \code{movedata_cols} for movement config files.
\item \code{holdingdata_fileopts} and \code{holdingdata_cols} for holding config files.
\item All of the above four top-level config keys for combined config files.
}
}
\keyword{internal}
