% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_config_type}
\alias{validate_config_type}
\title{Validate that the required config type is covered by the inferred config type}
\usage{
validate_config_type(required_config_type, inferred_config_type)
}
\arguments{
\item{required_config_type}{Character vector representing the type of
configurations that are required. One of \code{"movement"}, \code{"holding"}, or
\code{c("movement", "holding")} (containing both types of configurations). Within
the movenet app this is used to indicate the config type expected by the app page.}

\item{inferred_config_type}{Character vector representing the type of
configurations inferred to be present in the config file. One of \code{"movement"},
\code{"holding"}, or \code{c("movement", "holding")} (containing both types of
configurations).}
}
\value{
Either \code{NULL} if the required config type corresponds to or is a subset of the
inferred config type, or a character string stating that \code{file} has an unexpected config type.
}
\description{
Internal helper function that checks whether the required config type
corresponds to or is a subset of the inferred config type.
}
\details{
This is used by the movenet app to ensure that the config file uploaded
in the "movement data input" page contains movement configurations, and that the
config file uploaded in the "holding data input" page contains holding configurations.

N.B. \code{validate_config_type()} does not check config file validity, only whether
inferred and required config types correspond.
}
\keyword{internal}
