% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_config.R
\name{validate_yaml}
\alias{validate_yaml}
\title{Validate that a file can be parsed as a YAML string}
\usage{
validate_yaml(file)
}
\arguments{
\item{file}{Path of the file to validate.}
}
\value{
A named list consisting of 2 elements:
\itemize{
\item \code{test}: A boolean, whether \code{file} contains a valid YAML string that can be read by yaml.load_file.
\item \code{msg}: Either \code{NULL} if \code{isTRUE(test)}, or a character string stating that \code{file} is not valid YAML format.
}
}
\description{
Internal helper function that checks whether \code{file} has YAML format.
}
\details{
Adapted from https://rdrr.io/cran/validate/src/R/yaml.R
}
\keyword{internal}
