% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_helpers.R
\name{violinplot_monthly_measures}
\alias{violinplot_monthly_measures}
\title{Draw violin plot of distributions of monthly values of a selected network measure, for various jittered & rounded networks}
\usage{
violinplot_monthly_measures(monthly_data, measure_name)
}
\arguments{
\item{monthly_data}{tibble with measure values, for monthly networks}

\item{measure_name}{measure name (to refer to in y axis)}
}
\description{
Draw violin plot of distributions of monthly values of a selected network measure, for various jittered & rounded networks
}
