% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_creation.R
\name{are_ids_consec_intchars_from_1}
\alias{are_ids_consec_intchars_from_1}
\title{Are holding identifiers consecutive integer numbers, starting from 1?}
\usage{
are_ids_consec_intchars_from_1(holding_ids)
}
\arguments{
\item{holding_ids}{A character vector containing holding identifiers.}
}
\value{
A single boolean: \code{TRUE} if the identifiers are consecutive integer
numbers starting from 1; \code{FALSE} otherwise.
}
\description{
\code{are_ids_consec_intchars_from_1()} is an internal helper function that checks
whether holding identifiers are consecutive "integers" (in character format),
ranging from 1 to the total number of holding identifiers.
This is to facilitate conversion of movenet data tibbles into formats
compatible with \code{\link[networkDynamic:networkDynamic]{networkDynamic::networkDynamic()}} and \link[SimInf:SimInf]{SimInf::SimInf},
which require node identifiers to be consecutive integers starting from 1.
}
\keyword{internal}
