% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_contact_chains.R
\name{build_tree2}
\alias{build_tree2}
\title{Build contact tracing trees}
\usage{
build_tree2(network_structure)
}
\arguments{
\item{network_structure}{A data.frame with contact tracing data, resulting
from \code{EpiContactTrace::NetworkStructure()}.}
}
\value{
A list with two dataframes, \code{ingoing} and \code{outgoing}, containing
movement data for ingoing and outgoing contact chains, respectively. The
dataframes have columns "node" (holding id), "parent" (holding id of the
parent node) and "level" (distance from root). "level" is required for the
positioning of nodes in a schematic figure.
}
\description{
Internal helper function that processes contact tracing results from
EpiContactTrace's \code{Trace()} function into separate dataframes for ingoing
and outgoing contact chains.
}
\details{
Modified from \code{EpiContactTrace:::build_tree()} (fixed a bug in
parent assignment).
}
\keyword{internal}
