% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_config.R
\name{change_config}
\alias{change_config}
\alias{load_config}
\alias{save_config}
\alias{new_config}
\alias{get_config}
\title{Change configurations}
\usage{
load_config(configfile)

save_config(outfile)

new_config(config_type)

get_config(...)

change_config(...)
}
\arguments{
\item{configname}{some description}
}
\description{
Change configurations
}
