% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_config.R
\name{change_config}
\alias{change_config}
\alias{new_config}
\alias{get_config}
\title{Change configurations}
\usage{
change_config(country)

new_config()

get_config(parameter)
}
\arguments{
\item{country}{some description}
}
\description{
Change configurations
}
