% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_config.R
\name{change_config}
\alias{change_config}
\alias{load_config}
\alias{save_config}
\alias{new_config}
\alias{get_config}
\alias{movenet.options}
\alias{movenet.getOption}
\title{Change configurations}
\usage{
load_config(configname)

save_config(configname)

new_config()

get_config(...)

movenet.options(...)

movenet.getOption(name)
}
\arguments{
\item{configname}{some description}
}
\description{
Change configurations
}
