% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_config.R
\name{change_config}
\alias{change_config}
\title{Change configurations}
\usage{
change_config(surveillance_system)
}
\arguments{
\item{surveillance_system}{some description}
}
\description{
Change configurations
}
