% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_data.R
\name{colindex2name}
\alias{colindex2name}
\title{Replace \code{cols} configurations stored as column indices with the relevant
column headers}
\usage{
colindex2name(data, minvars, extra)
}
\arguments{
\item{data}{A tibble with ALL columns of a movement or holding data file.}

\item{minvars}{A named list with all required \code{cols} configurations.}

\item{extra}{A named list with all optional \code{cols} configurations set in the movenet environment.}
}
\description{
\code{colindex2name()} replaces c
}
\keyword{internal}
