% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_contact_chains.R
\name{contactchains2leaflet}
\alias{contactchains2leaflet}
\title{Create Leaflet map with ingoing and outgoing contact chains}
\usage{
contactchains2leaflet(
  movement_data,
  holding_data,
  tEnd = NULL,
  days = NULL,
  inBegin = NULL,
  inEnd = NULL,
  outBegin = NULL,
  outEnd = NULL,
  maxDistance = NULL,
  admin_areas_map = NULL,
  colour_domain_weight_in = NULL,
  colour_domain_weight_out = NULL,
  colour_domain_moves_in = NULL,
  colour_domain_moves_out = NULL
)
}
\arguments{
\item{movement_data}{A movenet-format movement data tibble, with movements
that are part of ingoing and outgoing contact chains.}

\item{holding_data}{A movenet-format holding data tibble, with holdings that
are part of ingoing and outgoing contact chains.}

\item{tEnd}{A single date, or character or factor that can be converted to a
date, representing the last date to include both ingoing and outgoing
movements. Defaults to \code{NULL}.}

\item{days}{A single numeric representing the number of days prior to \code{tEnd}
to include ingoing and outgoing movements for. Defaults to \code{NULL}.}

\item{inBegin}{A single date, or character or factor that can be converted to
a date, representing the first date to include ingoing movements for.
Defaults to \code{NULL}.}

\item{inEnd}{A single date, or character or factor that can be converted to a
date, representing the last date to include ingoing movements for. Defaults
to \code{NULL}.}

\item{outBegin}{A single date, or character or factor that can be converted
to a date, representing the first date to include outgoing movements for.
Defaults to \code{NULL}.}

\item{outEnd}{A single date, or character or factor that can be converted to
a date, representing the last date to include outgoing movements for.
Defaults to \code{NULL}.}

\item{maxDistance}{A single numeric representing the maximum 'distance' (in
number of movements) away from the root node, at which to stop contact
tracing. Defaults to \code{NULL}, i.e. don't use the maxDistance stop criterion.}

\item{admin_areas_map}{An sf object containing multipolygons for relevant
administrative areas.}

\item{colour_domain_weight_in}{A numeric vector representing the range (i.e.
\code{c(min, max)}) of possible aggregate ingoing weight values to colour
administrative areas by. If \code{NULL} (default), the range will be determined
from the data.}

\item{colour_domain_weight_out}{A numeric vector representing the range (i.e.
\code{c(min, max)}) of possible aggregate outgoing weight values to colour
administrative areas by.  If \code{NULL} (default), the range will be determined
from the data.}

\item{colour_domain_moves_in}{A numeric vector representing the range (i.e.
\code{c(min, max)}) of possible aggregate numbers of ingoing moves to colour
administrative areas by. If \code{NULL} (default), the range will be determined
from the data.}

\item{colour_domain_moves_out}{A numeric vector representing the range (i.e.
\code{c(min, max)}) of possible aggregate numbers of outgoing moves to colour
administrative areas by. If \code{NULL} (default), the range will be determined
from the data.}
}
\value{
A Leaflet map (HTML widget object) with graphics layers for holdings
and moves that are part of ingoing and outgoing contact chains, and for the
provided administrative area boundaries. Further layers can be added using
\verb{\%>\%}.
}
\description{
Internal helper function that plots ingoing and outgoing contact chains onto
a Leaflet map with default OpenStreetMap background.
}
\keyword{internal}
