% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_creation.R
\name{create_vertex_spells}
\alias{create_vertex_spells}
\title{Create a networkDynamic vertex.spells data frame from a movement data tibble}
\usage{
create_vertex_spells(movement_data)
}
\arguments{
\item{movement_data}{A movement data tibble with integer node identifiers and
dates in numeric format.}
}
\value{
A networkDynamic-compliant \code{vertex.spells} data frame, consisting of
3 columns indicating node activity onset (\code{date} from \code{movement_data}),
node activity terminus (again \code{date} from \code{movement_data}), and node
identifier (all identifiers in \code{from} and \code{to} in \code{movement_data}).
}
\description{
\code{create_vertex_spells()} is an internal helper function that converts a
modified movenet-format movement data tibble into a networkDynamic-compliant
node activity (\code{vertex.spells}) data frame. It requires integer node
identifiers, and dates converted to numeric format.
}
\seealso{
\code{\link[networkDynamic:networkDynamic]{networkDynamic::networkDynamic()}} for \code{vertex.spells} requirements
and how this data frame is used.
}
\keyword{internal}
