% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2contactmatrix.R
\name{data2contactmatrix}
\alias{data2contactmatrix}
\title{Title}
\usage{
data2contactmatrix(
  movement_data,
  holding_data = NULL,
  incl_nonactive_holdings = TRUE,
  accept_missing_coordinates = FALSE,
  accept_missing_additional_tm_probabilities = FALSE,
  weight_unit_transmission_probability = 0.5,
  whole_months = TRUE,
  local_spread_transmission_probabilities = NULL,
  additional_transmission_prob_matrices = NULL
)
}
\arguments{
\item{movement_data}{a movenet-format movement data tibble}

\item{holding_data}{a movenet-format holding data tibble (optional)}

\item{incl_nonactive_holdings}{whether to include non-active holdings
(default \code{FALSE})}

\item{weight_unit_transmission_probability}{transmission probability per unit
weight moved (default 1)}

\item{whole_months}{A logical indicating whether \code{movement_data} covers
full months (default \code{TRUE}). This affects calculation of daily average
weights, and thus of movement-based transmission probabilities. See
@details.}

\item{additional_transmission_prob_matrices}{named list with any additional
transmission probability matrices to include}

\item{local_spread_probabilities}{dataframe/tibble of tiered local spread
probabilities. 3 columns: lower_boundary (in metres), upper_boundary (in
metres), probability}
}
\value{
A named list with two elements:
\itemize{
\item \code{transmission_matrix} containing a matrix with overall daily transmission
probabilities from each holding (each row) to each holding (each column),
based on movements, (optionally) local spread, and (optionally) any other
user-provided transmission probabilities between holdings. Holdings have been
given new identifiers in the form of digits (format required by SimInf); the
original identifiers can be looked up in the provided \code{key}. The matrix
diagonal is set to 0, corresponding to no self-infection.
\item \code{key} containing a named character vector that links original holding
identifiers (element names) and numeric node identifiers (as element values).
}
}
\description{
Title
}
