% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2contactmatrix.R
\name{data2contactmatrix}
\alias{data2contactmatrix}
\title{Title}
\usage{
data2contactmatrix(
  movement_data,
  holding_data = NULL,
  incl_nonactive_holdings = TRUE,
  weight_unit_transmission_probability = 1,
  whole_months = TRUE,
  local_spread_transmission_probabilities = NULL,
  additional_transmission_prob_matrices = NULL
)
}
\arguments{
\item{movement_data}{a movenet-format movement data tibble}

\item{holding_data}{a movenet-format holding data tibble (optional)}

\item{incl_nonactive_holdings}{whether to include non-active holdings
(default \code{FALSE})}

\item{weight_unit_transmission_probability}{transmission probability per
unit weight moved (default 1)}

\item{whole_months}{whether the movement data covers full months (default
\code{TRUE}). This affects calculation of daily average weights, and thus of
movement-based transmission probabilities. See @details}

\item{additional_transmission_prob_matrices}{named list with any additional
transmission probability matrices to include}

\item{local_spread_probabilities}{dataframe/tibble of tiered local spread
probabilities. 3 columns: lower_boundary (in metres), upper_boundary (in
metres), probability}
}
\value{
a matrix of size number of holdings x number of holdings with overall
transmission probabilities, based on movements, (optionally) local spread,
and (optionally) any other user-provided transmission probabilities.
}
\description{
Title
}
