% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_creation.R
\name{holdingids2consecints}
\alias{holdingids2consecints}
\title{Replace holding identifiers with consecutive "integers" (in character format)
starting from 1, and harmonise movement and holding data}
\usage{
holdingids2consecints(
  movement_data,
  holding_data = NULL,
  incl_nonactive_holdings = FALSE
)
}
\arguments{
\item{movement_data}{A movenet-format movement data tibble.}

\item{holding_data}{A movenet-format holding data tibble (optional).}

\item{incl_nonactive_holdings}{A logical that indicates whether to keep entries in
\code{holding_data} for holdings that are not present in \code{movement_data}. Default
is \code{FALSE}.}
}
\value{
A named list with 3 elements:
\itemize{
\item \code{key}: A named character vector representing the identifier replacement key.
Names are original identifiers, values are new integer identifiers.
\item \code{movement_data}: \code{movement_data} with holding identifiers replaced with
integers.
\item \code{holding_data}: Either \code{holding_data} with holding identifiers replaced
with integers, and potentially some additional or removed rows; or \code{NULL}
if no \code{holding_data} was provided.
}
}
\description{
\code{holdingids2consecints()} is an internal helper function that replaces
holding identifiers in \code{movement_data} and (optional) \code{holding_data}
with consecutive "integers" (in character format), ranging from 1 to the
total number of node identifiers.

Additionally, if \code{holding_data} is provided, it checks whether the included
holding identifiers correspond to those in \code{movement_data}, and it harmonises
both data tibbles:
\itemize{
\item If \code{movement_data} contains additional holdings, entries for these are
added to \code{holding_data}, with any columns beyond \code{id} being assigned
the value \code{NA}.
\item If \code{holding_data} contains any entries for holdings that are not in
\code{movement_data} (i.e. entries for non-active holdings), these are removed if
\code{incl_nonactive_holdings == FALSE} or kept otherwise.
}

This function facilitates conversion of movenet data tibbles into formats
compatible with \code{\link[networkDynamic:networkDynamic]{networkDynamic::networkDynamic()}} and \link[SimInf:SimInf]{SimInf::SimInf},
which require node identifiers to be consecutive integers starting from 1.
}
\keyword{internal}
