% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsening_functions.R
\name{jitter_dates}
\alias{jitter_dates}
\title{Jitter (add noise to) movement dates}
\usage{
jitter_dates(data, range)
}
\arguments{
\item{data}{A movenet-format movement tibble.}

\item{range}{A positive integer, indicating the maximum amount of jitter (in
days) to apply to the movement dates (see Details).}
}
\value{
A movement tibble like \code{data}, but with jittered movement dates.
}
\description{
\code{jitter_dates()} adds random noise (jitter) to movement dates in a
movenet-format movement tibble.
}
\details{
Requires that an appropriate movement config file is loaded, to correctly
identify the \code{date} column in \code{data}.

Movement dates are modified by addition of a number of days between \code{-range}
and \code{range}, following a discrete uniform distribution. If this were to
result in a date being moved beyond the original date range, the amount of
jitter for this date is resampled, until the resulting date is within the
original date range.
}
\examples{
# Set-up: Save movenet environment with current configurations
movenetenv <- movenet:::movenetenv
old_config <- movenetenv$options

# Load a movement config file
load_config(system.file("configurations", "ScotEID.yml",
                        package = "movenet"))

# Add -3 to +3 days of jitter to movement dates
head(example_movement_data)
moves_w_jittered_dates <- jitter_dates(example_movement_data, 3)
head(moves_w_jittered_dates)

# Clean-up: Reinstate previous configurations
movenetenv$options <- old_config
rm("old_config", "movenetenv", "moves_w_jittered_dates")

}
\seealso{
Other Privacy-enhancing functions: 
\code{\link{anonymise}()},
\code{\link{create_anonymisation_effect_analysis_report}()},
\code{\link{jitter_weights}()},
\code{\link{round_dates}()},
\code{\link{round_weights}()}
}
\concept{Privacy-enhancing functions}
