% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsening_functions.R
\name{jitter_weights}
\alias{jitter_weights}
\title{Jitter (add noise to) numeric movement data}
\usage{
jitter_weights(data, range, column = movenetenv$options$movedata_cols$weight)
}
\arguments{
\item{data}{A movenet-format movement tibble.}

\item{range}{A positive number, indicating the maximum amount of jitter to
apply to the numeric data (see Details).}

\item{column}{Name of a single numeric column to jitter. By default this is
the weight column (as specified in the loaded movement configurations).}
}
\value{
A movement tibble like \code{data}, but with jitter applied to the selected
numeric column.
}
\description{
\code{jitter_weights()} adds random noise (jitter) to a selected numeric column
in a movenet-format movement tibble. By default, the weight column is
selected.
}
\details{
Requires that an appropriate movement config file is loaded, to correctly
identify the \code{weight} column in \code{data}.

The data in the selected column are modified by addition of an amount of
noise between \code{-range} and \code{range}, following a uniform distribution.
If this were to result in a data point becoming \verb{<= 0}, the amount of jitter
for this data point is resampled, until the resulting data point becomes
positive. This is to capture that any movement in a livestock movement
database is assumed to have a positive weight (quantity of animals moved).
}
\examples{
# Set-up: Save movenet environment with current configurations
movenetenv <- movenet:::movenetenv
old_config <- movenetenv$options

# Load a movement config file
load_config(system.file("configurations", "ScotEID.yml",
                        package = "movenet"))

# Add random noise from -3 to +3 to movement weights
head(example_movement_data)
moves_w_jittered_weights <- jitter_weights(example_movement_data, 3)
head(moves_w_jittered_weights)

# Add random noise from -3 to +3 to movement reference numbers
moves_w_jittered_refs <- jitter_weights(example_movement_data,
                                        range = 3,
                                        column = "movement_reference")
head(moves_w_jittered_refs)

# Clean-up: Reinstate previous configurations
movenetenv$options <- old_config
rm("old_config", "movenetenv", "moves_w_jittered_weights",
   "moves_w_jittered_refs")

}
\seealso{
\code{\link[base:jitter]{base::jitter()}}, which this function is based on.

Other Privacy-enhancing functions: 
\code{\link{anonymise}()},
\code{\link{jitter_dates}()},
\code{\link{round_dates}()},
\code{\link{round_weights}()}
}
\concept{Privacy-enhancing functions}
