% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations-related_functions.R
\name{load_config}
\alias{load_config}
\title{Load configurations from a config file into the movenet environment}
\usage{
load_config(configfile)
}
\arguments{
\item{configfile}{Either the path to a YAML-format movenet config file, or the name of a
pre-installed config file in the movenet package's \code{configurations} folder.}
}
\value{
No return value, but \code{load_config()} has the effect of copying the
provided configurations into the movenet environment. If this is
successful, a message is printed.
}
\description{
\code{load_config()} checks that the provided \code{configfile} is valid, and then loads
its content into the movenet environment. This lets the movenet package know
how to read and interpret upcoming input data files.
}
\details{
Values of configurations representing column headers or indices (in
the config file \code{fileopts} section) are converted to ASCII-compliant and
syntactically valid names before being loaded into in the environment.
}
\examples{
# Set-up: Save movenet environment with current configurations
movenetenv <- movenet:::movenetenv
old_config <- movenetenv$options

# Load a config file using a path
load_config(system.file("configurations", "fakeScotEID_holding.yml", package="movenet"))
get_config() #Examine configurations

# Alternatively, load a pre-installed config file by using just its name
load_config("ScotEID.yml")
get_config() #Examine configurations

# Clean-up: Reinstate previous configurations
movenetenv$options <- old_config
rm("old_config", "movenetenv")

}
\seealso{
\itemize{
\item \code{\link[=validate_config]{validate_config()}} for the underlying config file validation process.
\item \code{\link[=asciify]{asciify()}} for the underlying ASCIIfication process.
\item \code{vignette("configurations")} for an explanation of the movenet config system.
\item \code{list.files(system.file("configurations", package = "movenet"))} for a list of pre-installed templates and validated config files.
}

Other configurations-related functions: 
\code{\link{change_config}()},
\code{\link{get_config}()},
\code{\link{new_config}()},
\code{\link{save_config}()},
\code{\link{validate_config}()}
}
\concept{configurations-related functions}
