% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_network_helpers.R
\name{movedata2networkDynamic}
\alias{movedata2networkDynamic}
\title{Create dynamic networks from movenet-format, anonymised movement data}
\usage{
movedata2networkDynamic(data)
}
\arguments{
\item{data}{Movenet format movement data, with integerish holding IDs}
}
\description{
Create dynamic networks from movenet-format, anonymised movement data
}
