% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movenet-package.R
\docType{package}
\name{movenet-package}
\alias{movenet}
\alias{movenet-package}
\title{movenet: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

}
\keyword{internal}
