% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movenet-package.R
\docType{package}
\name{movenet-package}
\alias{movenet}
\alias{movenet-package}
\title{movenet: A Toolkit Facilitating the Use of Livestock Movement Data in Veterinary Public Health}
\description{
A set of workflows facilitating the dataflow from livestock movement data to social network analysis and disease transmission models. 'movenet' also addresses common data issues such as the diversity of data formats and privacy preservation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://digivet-consortium.github.io/movenet/}
}

}
\author{
\strong{Maintainer}: Carlijn Bogaardt \email{Carlijn.Bogaardt@glasgow.ac.uk} (\href{https://orcid.org/0000-0002-9322-6978}{ORCID})

Authors:
\itemize{
  \item Matt Denwood \email{md@sund.ku.dk}
  \item Stefan Widgren \email{stefan.widgren@sva.se} (\href{https://orcid.org/0000-0001-5745-2284}{ORCID})
  \item Jessica Enright \email{Jessica.Enright@glasgow.ac.uk} (\href{https://orcid.org/0000-0002-0266-3292}{ORCID})
}

Other contributors:
\itemize{
  \item Ruby Graham [contributor]
  \item Petter Hopp [contributor]
  \item Thomas Rosendal [contributor]
}

}
\keyword{internal}
