% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_config.R
\name{new_config}
\alias{new_config}
\title{Copying a config file template to your working directory}
\usage{
new_config(config_type = c("movement", "holding"))
}
\arguments{
\item{config_type}{}
}
\description{
Copying a config file template to your working directory
}
