% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{parallel_summarise_temporal_node_properties}
\alias{parallel_summarise_temporal_node_properties}
\title{Extract summary stats for temporal network node properties, in parallel}
\usage{
parallel_summarise_temporal_node_properties(
  networks,
  n_threads,
  node_property,
  statistics,
  identify_nodes = FALSE
)
}
\arguments{
\item{networks}{A list containing movement networks (networkDynamic objects)}

\item{n_threads}{Numeric indicating number of threads over which to
parallelise.}

\item{node_property}{Character indicating the network metric for which to
calculate summary statistics. One of \code{"forward reachability"},
\code{"temporal degree"}, \code{"temporal indegree"}, or \code{"temporal outdegree"}.}

\item{statistics}{A list with summary function(s) to calculate for the
selected \code{node_property}. For example, \code{list(median, max = max)}.}

\item{identify_nodes}{A logical indicating whether you want to identify the
nodes with maximal values for \code{node_property} in each network. Default is
\code{FALSE}. Requires that \code{max} is included as a named function within
\code{statistics}.}
}
\value{
A named list of lists. For each network, a list consisting of: (1) a
named list \code{summary_statistics}, with the selected node property summary
statistic(s), (2 - only if \code{identify_nodes} is set as \code{TRUE}) a named
character vector \code{node_pid_with_max_value}, with the persistent
identifier(s) of the node(s) with maximal value(s) for the selected node
property.
}
\description{
\code{parallel_summarise_temporal_node_properties()} calculates summary statistics
for a selected node property, over a list of networks. Optionally, identifiers
of the nodes with maximum values of the selected node property can also be
returned. This function uses a socket cluster with \code{n_threads} threads to
run calculations in parallel.
}
