% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_data.R
\name{reformat_coords}
\alias{reformat_coords}
\title{Check coordinate columns & change to sf geometry}
\usage{
reformat_coords(coord_x_col, coord_y_col, decimal, crs)
}
\arguments{
\item{crs}{}
}
\value{
a tibble with sf geometry point data and attributes including crs
}
\description{
Check coordinate columns & change to sf geometry
}
\keyword{internal}
