% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_movement_data.R
\name{reformat_move_data}
\alias{reformat_move_data}
\title{Reading and reformatting of animal movement data}
\usage{
reformat_move_data(move_data_file)
}
\arguments{
\item{move_data_file}{Path to a delimited file with movement data (alternatively: literal data or a connection).}
}
\value{
reformatted movement data (selected & renamed columns)
}
\description{
Reformats movement data from a delimited file into a common intermediate format (extracts and renames selected columns)
}
