% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_config.R
\name{save_config}
\alias{save_config}
\title{Saving configurations to a movenet config file}
\usage{
save_config(outfile, config_type = c("movement", "holding"))
}
\arguments{
\item{outfile}{}

\item{config_type}{}
}
\description{
Saving configurations to a movenet config file
}
