% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRcm.R
\name{SEIRcm}
\alias{SEIRcm}
\title{Create an SEIR model with contact matrix}
\usage{
SEIRcm(u0 = NULL, tspan = NULL, beta = NULL, epsilon = NULL, gamma = NULL)
}
\arguments{
\item{u0}{FIXME}

\item{tspan}{FIXME}

\item{beta}{the transmission rate from susceptible to infected
state.}

\item{epsilon}{the incubation rate from exposed to infected state.}

\item{gamma}{the recovery rate from infected to recovered state.}

\item{coupling}{a measure of the strength of the interaction
between nodes. Specifically, \code{coupling[i, j]} measures
the relative strength of transmission to node \code{i} from
node \code{j}.}
}
\description{
Create an SEIR model with contact matrix
}
