% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_data.R
\name{select_cols}
\alias{select_cols}
\title{Check for the presence of required and additional columns, and extract from \code{data}}
\usage{
select_cols(data, minvars, extra)
}
\arguments{
\item{data}{A tibble with ALL columns of a movement or holding data file.}

\item{minvars}{A named list with all required \code{cols} configurations.}

\item{extra}{A named list with all optional \code{cols} configurations set in the movenet environment.}
}
\description{
Check for the presence of required and additional columns, and extract from \code{data}
}
\keyword{internal}
